Check that distribution is sums up to one.

Check if the given .data is a distribution and normalise it if necessary with an optional laplace correction.

check_distribution(.data, .do.norm = NA, .laplace = 1, .na.val = 0,
  .warn.zero = F, .warn.sum = T)

Arguments

.data
Numeric vector of values.

.do.norm
One of the three values - NA, T or F. If NA than check for distrubution (sum(.data) == 1) and normalise if needed with the given laplace correction value. if T then do normalisation and laplace correction. If F than don't do normalisaton and laplace correction.

.laplace
Value for the laplace correction.

.na.val
Replace all NAs with this value.

.warn.zero
if T then the function checks if in the resulted vector (after normalisation) are any zeros, and prints a warning message if there are some.

.warn.sum
if T then the function checks if the sum of resulted vector (after normalisation) is equal to one, and prints a warning message if not.

Value

Numeric vector.