Clustering of objects or distance matrices

Cluster the data with one of the following methods:

- immunr_hclust clusters the data using the hierarchical clustering from hcut;

- immunr_kmeans clusters the data using the K-means algorithm from kmeans;

- immunr_dbscan clusters the data using the DBSCAN algorithm from dbscan.

immunr_hclust(.data, .k = 2, .k.max = nrow(.data) - 1, .method = "complete", .dist = T)

immunr_kmeans(.data, .k = 2, .k.max = as.integer(sqrt(nrow(.data))) + 1,
.method = c("silhouette", "gap_stat"))

immunr_dbscan(.data, .eps, .dist = T)

Arguments

.data
Matrix or data frame with features, distance matrix or output from repOverlapAnalysis or geneUsageAnalysis functions.

.k
The number of clusters to create, passed as k to hcut or as centers to kmeans.

.k.max
Limits the maximum number of clusters. It is passed as k.max to fviz_nbclust for immunr_hclust and immunr_kmeans.

.method
Passed to hcut or as fviz_nbclust.

In case of hcut the agglomeration method is going to be used (argument hc_method).

In case of fviz_nbclust it is the method to be used for estimating the optimal number of clusters (argument method).

.dist
If TRUE then ".data" is expected to be a distance matrix. If FALSE then the euclidean distance is computed for the input objects.

.eps
Local radius for expanding clusters, minimal distance between points to expand clusters. Passed as eps to dbscan.