Dimensionality reduction

Collect a set of principal variables, reducing the number of not important variables to analyse. Dimensionality reduction makes data analysis algorithms work faster and sometimes more accurate, since it also reduces noise in the data. Currently available methods are:

- immunr_pca performs PCA (Principal Component Analysis) using prcomp;

- immunr_mds performs MDS (Multi-Dimensional Scaling) using isoMDS;

- immunr_tsne performs tSNE (t-Distributed Stochastic Neighbour Embedding) using Rtsne.

immunr_pca(.data, .scale = default_scale_fun, .raw = TRUE, .orig = FALSE, .dist = FALSE)

immunr_mds(.data, .scale = default_scale_fun, .raw = TRUE, .orig = FALSE, .dist = TRUE)

immunr_tsne(.data, .perp = 1, .dist = TRUE, ...)

Arguments

.data
A matrix or a data frame with features, distance matrix or output from repOverlapAnalysis or geneUsageAnalysis functions.

.scale
A function to apply to your data before passing it to any of dimensionality reduction algorithms. There is no scaling by default.

.raw
If TRUE then return non-processed output from dimensionality reduction algorithms. Pass FALSE if you want to visualise results.

.orig
If TRUE then return the original result from algorithms. Pass FALSE if you want to visualise results.

.dist
If TRUE then assume ".data" is a distance matrix.

.perp
The perplexity parameter for Rtsne. Sepcifies the number of neighbours each data point must have in the resulting plot.

...
Other parameters passed to Rtsne.

Value

immunr_pca - an output from prcomp.

immunr_mds - an output from isoMDS.

immunr_tsne - an output from Rtsne.

See also

Examples

data(immdata)
gu <- geneUsage(immdata$data)
gu[is.na(gu)] <- 0
gu <- t(as.matrix(gu[, -1]))
immunr_pca(gu)

#> Standard deviations (1, .., p=12):
#>  [1] 3.630914e+00 3.232821e+00 2.359621e+00 2.029413e+00 1.848420e+00
#>  [6] 1.687546e+00 1.556702e+00 1.492562e+00 1.261700e+00 1.147413e+00
#> [11] 9.246792e-01 7.931205e-16
#> 
#> Rotation (n x k) = (48 x 12):
#>                 PC1          PC2         PC3          PC4          PC5
#>  [1,] -0.0642668458  0.134479401 -0.05234665 -0.351917954  0.016668727
#>  [2,] -0.1483986328  0.115476723  0.01674999 -0.304057276  0.032936023
#>  [3,]  0.0046965924  0.209919426  0.03950533 -0.249520773  0.063521543
#>  [4,] -0.1193468206  0.037588388 -0.05169612  0.113385276  0.057553505
#>  [5,] -0.1861215511 -0.059720978 -0.12410548 -0.113426166  0.023369616
#>  [6,] -0.1211820748  0.178388782 -0.17885121  0.099057097 -0.048658383
#>  [7,] -0.0815799131  0.236039396  0.08026686  0.067465032  0.022695054
#>  [8,] -0.2104800655 -0.004328549 -0.13554495 -0.201756284  0.172163224
#>  [9,] -0.0924076776 -0.015658812 -0.36148366  0.105294591  0.064124315
#> [10,] -0.2028578175 -0.163952406  0.07312632  0.020470706  0.153491979
#> [11,] -0.1648029305 -0.050522817  0.25522864 -0.028374260 -0.182905550
#> [12,] -0.1704803313 -0.103670921 -0.09029589  0.061022978  0.176425431
#> [13,] -0.1968070236 -0.138492263  0.16477036  0.139174931  0.030364702
#> [14,] -0.2413454583 -0.096460212  0.01381341 -0.028833400  0.014448306
#> [15,] -0.1713955212 -0.147412155 -0.02668257  0.002248364  0.239129982
#> [16,] -0.1078882679  0.071540857  0.17852513  0.175737545 -0.188398997
#> [17,] -0.1484251756 -0.171454469 -0.01210451 -0.058337561  0.148109245
#> [18,] -0.1917958754 -0.136370286 -0.14073506  0.073670277 -0.067096030
#> [19,] -0.1476591568  0.222509877 -0.09738207  0.111748730  0.059866585
#> [20,]  0.0423067467 -0.127684655 -0.14580417 -0.110678315  0.068413945
#> [21,] -0.0903189244 -0.024919730 -0.11767844 -0.016783898 -0.123800887
#> [22,] -0.0782342048 -0.135554583  0.07139625  0.380602367  0.101213239
#> [23,]  0.1250223338 -0.049347354  0.06539111  0.196075034 -0.003347335
#> [24,] -0.2056831200 -0.063200796  0.01122275  0.019001660  0.274672332
#> [25,] -0.2233864705 -0.053958823  0.15716853 -0.151290701  0.055789865
#> [26,]  0.0410018350  0.233749392  0.08218583  0.073393505  0.118769294
#> [27,]  0.1247869497  0.205414886 -0.12227693 -0.001227982  0.130105664
#> [28,]  0.0349674392 -0.157233243 -0.29755249  0.041665365 -0.234730097
#> [29,] -0.1311146496  0.126245864  0.28055263  0.136246368  0.078269384
#> [30,] -0.1825931608  0.104444741  0.06451978 -0.044114279 -0.206358095
#> [31,] -0.1870934972  0.149403918  0.08326795 -0.007589653 -0.023693516
#> [32,] -0.1517349610  0.188275956  0.09706919  0.143804022 -0.190812606
#> [33,] -0.1507532011  0.044049365 -0.03419754 -0.009296996 -0.247089325
#> [34,] -0.1693491427  0.139879866  0.16183081  0.014309427  0.035763767
#> [35,]  0.0071780486 -0.151635168 -0.09175489 -0.051074531 -0.353763789
#> [36,]  0.0074043103  0.225374692 -0.10738489 -0.048283994 -0.116366317
#> [37,]  0.0007565336  0.275466390  0.03880796 -0.091183998 -0.172292790
#> [38,] -0.2261026196  0.062493040 -0.09255325 -0.084628028  0.026307890
#> [39,] -0.1184936274 -0.084675273  0.05907628 -0.291831192 -0.025711762
#> [40,] -0.1777712583 -0.138863105  0.16601182  0.104648694 -0.125108678
#> [41,]  0.0615043217 -0.218658302 -0.02980713  0.008774701 -0.310977967
#> [42,] -0.0442672766  0.226026448 -0.16542605  0.248268808  0.010082777
#> [43,] -0.0829976397 -0.112363133  0.04352047 -0.264277079 -0.196285006
#> [44,] -0.1205182577  0.216747245 -0.10756026  0.043071191 -0.060083884
#> [45,] -0.1276787483  0.153947409 -0.26118945  0.054632985 -0.030526648
#> [46,] -0.2096306602 -0.007892979 -0.02601878  0.046058616 -0.253622053
#> [47,] -0.1994368462 -0.053266997 -0.12994436  0.169301623 -0.092786147
#> [48,] -0.0915019380 -0.035574139 -0.35659066  0.027336791  0.031333950
#>                PC6           PC7          PC8          PC9         PC10
#>  [1,] -0.130423346 -0.0035551348 -0.055511515  0.235262131 -0.186705027
#>  [2,] -0.011641790  0.1178976198 -0.103332107  0.268541101 -0.091839830
#>  [3,]  0.090763231  0.0101691718  0.314440390  0.030030374 -0.066506971
#>  [4,] -0.246831338 -0.0191298596  0.419793250 -0.144294916  0.105994593
#>  [5,] -0.245448549  0.1114690339 -0.060281369  0.234019880  0.181329496
#>  [6,]  0.268607319 -0.0429757244 -0.049301186  0.124657951  0.053453420
#>  [7,]  0.133157076  0.0071032718  0.076729204 -0.036028656  0.334389135
#>  [8,]  0.069810223 -0.0138839954 -0.080728089 -0.010974244  0.097366679
#>  [9,]  0.120322264 -0.0887887701 -0.056972548 -0.043293589 -0.053353256
#> [10,]  0.076549688 -0.1148407803 -0.075808755  0.031168880  0.002801099
#> [11,] -0.005608586  0.0065916805 -0.208680625  0.046289024 -0.158706996
#> [12,] -0.207539893 -0.2242631619  0.149889417  0.071050782 -0.161292234
#> [13,] -0.082913080  0.0924585309  0.035467192  0.013304308  0.061197296
#> [14,] -0.012927593 -0.1469775186 -0.081509615  0.152731852 -0.135602803
#> [15,]  0.007550533  0.1383315526  0.014417525  0.137492108  0.286599756
#> [16,]  0.047181368 -0.2501920982  0.193049495  0.167475243  0.017188779
#> [17,]  0.298276784  0.0380493457  0.045587914  0.042736334 -0.171884105
#> [18,]  0.016082930  0.0607709495  0.146858827 -0.139083116 -0.214534059
#> [19,] -0.138107798  0.0089752253  0.089294389 -0.006613862 -0.044539473
#> [20,]  0.247943436 -0.0915950102  0.338651474  0.218550811  0.257341204
#> [21,] -0.188111467 -0.4167770222  0.107358290  0.331823267 -0.098949699
#> [22,] -0.031191915  0.0009625946 -0.038367490  0.123510316 -0.113533288
#> [23,]  0.315371757  0.0712787977  0.032429839  0.346328841 -0.222443279
#> [24,]  0.098630352 -0.0599485949 -0.040726978 -0.006876107  0.270755946
#> [25,] -0.003325884  0.0493288820  0.127103633 -0.041202697  0.001855504
#> [26,]  0.003960727 -0.2739457700  0.027089823  0.034824136  0.113027514
#> [27,]  0.163117076 -0.1282125485 -0.093056581 -0.118730454  0.015889430
#> [28,]  0.044683658  0.0163940714 -0.028096963  0.022586405  0.101409954
#> [29,]  0.009368238  0.0434447418 -0.113872641  0.126821369  0.043652263
#> [30,] -0.286740381  0.0153730942  0.027193882 -0.128364928  0.069980629
#> [31,]  0.077173951  0.2574673094 -0.172915893  0.051323685  0.088773967
#> [32,] -0.001531095 -0.1546327460  0.038039483  0.016634327  0.016691997
#> [33,] -0.007189393 -0.2073594561 -0.338429869 -0.016350598  0.201510112
#> [34,]  0.255103255 -0.1078649960 -0.049431049 -0.084047763  0.092115170
#> [35,]  0.046844704  0.2568947281  0.099954744  0.197199840  0.111074980
#> [36,]  0.187801594  0.0989818329  0.241254889  0.206283625 -0.102271993
#> [37,]  0.019491931 -0.0478144453  0.007486442  0.148251470  0.126565956
#> [38,] -0.085769193  0.2622839074  0.054999507 -0.025050376 -0.045720174
#> [39,]  0.246802361 -0.1390823272 -0.076976788 -0.231837826 -0.186875698
#> [40,]  0.078845712 -0.0188847291  0.180676001 -0.055725850 -0.149477852
#> [41,] -0.010468419 -0.0188996895 -0.072999232  0.096980356  0.241325052
#> [42,]  0.028793434  0.0754524537 -0.058615296 -0.054857867 -0.061409526
#> [43,]  0.095935188 -0.2003357096  0.160951950 -0.300237808  0.107606881
#> [44,] -0.038607339  0.1980826198  0.150859648 -0.169000434 -0.118289335
#> [45,] -0.055630196  0.0183282949 -0.187578625 -0.009543368 -0.196877415
#> [46,]  0.216486199 -0.0066369278  0.021676041 -0.156545843 -0.102625786
#> [47,]  0.121024515  0.2300809801  0.049579785 -0.053324673  0.108577212
#> [48,]  0.002590157 -0.2129941290 -0.128523278 -0.066976463  0.044014428
#>               PC11         PC12
#>  [1,] -0.218805694 -0.050920759
#>  [2,]  0.024618249  0.007000995
#>  [3,] -0.102813737  0.054883716
#>  [4,] -0.347737581  0.029450891
#>  [5,] -0.171610496 -0.109162272
#>  [6,]  0.062637477 -0.195706294
#>  [7,]  0.266203573  0.087242313
#>  [8,]  0.059241795  0.295262049
#>  [9,]  0.165098716 -0.162533862
#> [10,] -0.042544587 -0.051743172
#> [11,]  0.036888852 -0.093186799
#> [12,]  0.025509937 -0.029033142
#> [13,] -0.087280907  0.002089324
#> [14,]  0.028165829  0.051334039
#> [15,] -0.034479742 -0.084596549
#> [16,]  0.307871968 -0.067695093
#> [17,] -0.103998500  0.104980590
#> [18,] -0.165938884  0.065640722
#> [19,]  0.065592673 -0.022777207
#> [20,]  0.091077144 -0.083527623
#> [21,]  0.152096016  0.301898583
#> [22,] -0.152863842 -0.009530624
#> [23,] -0.194426216  0.021941373
#> [24,]  0.067210283  0.282501180
#> [25,]  0.169720148 -0.079728317
#> [26,] -0.335983554 -0.172388640
#> [27,] -0.253528676  0.386235709
#> [28,]  0.131098053  0.052015464
#> [29,]  0.104535987  0.108870276
#> [30,] -0.028968490  0.011239025
#> [31,] -0.081640766  0.186945004
#> [32,] -0.048200371  0.048438248
#> [33,] -0.244837513 -0.054581759
#> [34,] -0.171004437 -0.034099985
#> [35,] -0.091920467  0.183415133
#> [36,] -0.123287806 -0.024775816
#> [37,] -0.008815415 -0.208478123
#> [38,]  0.093709903 -0.187241932
#> [39,]  0.053352123 -0.120593382
#> [40,] -0.069726564  0.110151641
#> [41,] -0.108532337  0.211585408
#> [42,]  0.010954379  0.095548005
#> [43,] -0.093553371  0.063971792
#> [44,]  0.149929420  0.177914227
#> [45,]  0.058177686  0.193710259
#> [46,] -0.016074760 -0.047589064
#> [47,] -0.102433681 -0.197711633
#> [48,] -0.080412934 -0.227537747

immunr_mds(dist(gu))

#> Standard deviations (1, .., p=4):
#> [1] 0 0 0 0
#> 
#> Rotation (n x k) = (12 x 2):
#>               [,1]       [,2]
#> A2-i129  252.18871 -133.61559
#> A2-i131 -164.42727   61.41270
#> A2-i133 -413.26445 -328.05184
#> A2-i132 -376.58882  -94.45000
#> A4-i191 -215.34836  246.11579
#> A4-i192 -326.65647  151.37319
#> MS1      -85.43483   14.39255
#> MS2      435.65357 -472.35259
#> MS3      -66.37501  160.34716
#> MS4      395.49211 -143.62405
#> MS5      -82.89849  135.73563
#> MS6      647.65931  402.71706

immunr_tsne(dist(gu))

#>            DimI      DimII
#>  [1,] -488.6956 -127.24825
#>  [2,]  322.4858  158.75745
#>  [3,] -134.4574  -49.72472
#>  [4,] -137.4079  -14.12452
#>  [5,]  504.6413   41.32553
#>  [6,]  344.2645   16.51346
#>  [7,]  450.9079   55.80497
#>  [8,] -495.8587  -47.12040
#>  [9,]  354.1779  108.04023
#> [10,] -515.6976 -100.62247
#> [11,]  373.0709   72.85408
#> [12,] -577.4311 -114.45536
#> attr(,"class")
#> [1] "matrix"      "immunr_tsne"