Incremental counting of repertoire similarity

Like in paper https://www.pnas.org/content/111/16/5980 (Fig. 4).

inc_overlap(.data, .fun, .step = 1000, .n.steps = 10,
  .downsample = F, .bootstrap = NA, .verbose.inc = T, ...)

Arguments

.data
The data to be processed. Can be data.frame, data.table, or a list of these objects.

Every object must have columns in the immunarch compatible format. immunarch_data_format

Competent users may provide advanced data representations: DBI database connections, Apache Spark DataFrame from copy_to or a list of these objects. They are supported with the same limitations as basic objects.

Note: each connection must represent a separate repertoire.

.fun
Function to compute overlaps. e.g., morisita_index.

.step
Either an integer or a numeric vector.

In the first case, the integer defines the step of incremental overlap.

In the second case, the vector encodes all repertoire sampling depths.

.n.steps
Integer. Number of steps if .step is a single integer. Skipped if ".step" is a numeric vector.

.downsample
If T than perform downsampling to N clonotypes at each step instead of choosing the top N clonotypes.

.bootstrap
Pass NA to turn off any bootstrapping, pass a number to perform bootstrapping with this number of tries.

.verbose.inc
Logical. If TRUE than show output from the computation process.

...
Other arguments passed to .fun.