Analysis of overlaps or any distance matrices

The repOverlapAnalysis function contains advanced data analysis methods. You can use several clustering and dimensionality reduction techniques in order to investigate further the difference between repertoires provided.

To cluster a subset of similar data with repOverlapAnalysis you can perform hierarchical clustering, k-means or dbscan ('hclust', 'kmeans', 'dbscan' respectively).

To reduce dimensions, for example, to select features for subsequent analysis, you can execute the multidimensional scaling or t-sne algorithms ('mds' and 'tsne' respectively).

repOverlapAnalysis(.data, .analysis = ("hclust"),
  .scale = default_scale_fun, .raw = T, .perp = 1, .theta = 0.1,
  .eps = 0.01, .k = 2)

Arguments

.data
Any distance matrix between pairs of repertoires. You can also pass your output from repOverlap.

.analysis
A string that defines the type of analysis to perform.

.scale
A function to scale the data before passing it to the MDS algorithm.

.raw
A logical value. Pass TRUE if you want to receive raw output of clustering or dimensionality reduction function of choice. Pass FALSE if you want to receive processed output that can be subjected to visualisation with vis function.

.perp
A numerical value, t-SNE parameter, see immunr_tsne.

.theta
A numerical value, t-SNE parameter, see immunr_tsne.

.eps
A numerical value, DBscan epsylon parameter, see immunr_dbscan.

.k
The number of clusters to create, passed as k to hcut or as centers to kmeans.