Save immune repertoires to the disk

The repSave function is deigned to save your data to the disk in desirable format. Currently supports "immunarch" and "vdjtools" file formats.

Usage

repSave(.data, .path, .format = c("immunarch", "vdjtools"), .compress = TRUE)

Arguments

.data

An R dataframe, a list of R dataframes or a list with data and meta where first element is a list of dataframes and the latter is a dataframe with metadata.

.path

A string with the path to the output directory. It should include file name if a single dataframe is provided to .data argument.

.format

A string with desirable format specification. Current options are "immunarch" and "vdjtools".

.compress

A boolean value. Defines whether the output will be compressed or not.

Value

No return value.

Details

It is not necessary to create directories beforehand. If the provided directory does not exist it will be created automatically.

Examples

# \dontrun{
data(immdata)
# Reduce data to save time on examples
immdata$data <- map(immdata$data, ~ .x %>% head(10))
#> Error in map(immdata$data, ~.x %>% head(10)): could not find function "map"
dirpath <- tempdir()
# Save the list of repertoires
repSave(immdata, dirpath)
#> Warning: The `path` argument of `write_tsv()` is deprecated as of readr 1.4.0.
#> ℹ Please use the `file` argument instead.
#> ℹ The deprecated feature was likely used in the immunarch package.
#>   Please report the issue at <https://github.com/immunomind/immunarch/issues>.
#> Warning: The `path` argument of `write_lines()` is deprecated as of readr 1.4.0.
#> ℹ Please use the `file` argument instead.
#> ℹ The deprecated feature was likely used in the immunarch package.
#>   Please report the issue at <https://github.com/immunomind/immunarch/issues>.
# Load it and check if it is the same
new_immdata <- repLoad(dirpath)
#> 
#> == Step 1/3: loading repertoire files... ==
#> Processing "/tmp/RtmpVH5iE5" ...
#>   -- [1/18] Parsing "/tmp/RtmpVH5iE5/A2-i129.tsv.gz" -- 
#> immunarch
#>   -- [2/18] Parsing "/tmp/RtmpVH5iE5/A2-i131.tsv.gz" -- 
#> immunarch
#>   -- [3/18] Parsing "/tmp/RtmpVH5iE5/A2-i132.tsv.gz" -- 
#> immunarch
#>   -- [4/18] Parsing "/tmp/RtmpVH5iE5/A2-i133.tsv.gz" -- 
#> immunarch
#>   -- [5/18] Parsing "/tmp/RtmpVH5iE5/A4-i191.tsv.gz" -- 
#> immunarch
#>   -- [6/18] Parsing "/tmp/RtmpVH5iE5/A4-i192.tsv.gz" -- 
#> immunarch
#>   -- [7/18] Parsing "/tmp/RtmpVH5iE5/MS1.tsv.gz" -- 
#> immunarch
#>   -- [8/18] Parsing "/tmp/RtmpVH5iE5/MS2.tsv.gz" -- 
#> immunarch
#>   -- [9/18] Parsing "/tmp/RtmpVH5iE5/MS3.tsv.gz" -- 
#> immunarch
#>   -- [10/18] Parsing "/tmp/RtmpVH5iE5/MS4.tsv.gz" -- 
#> immunarch
#>   -- [11/18] Parsing "/tmp/RtmpVH5iE5/MS5.tsv.gz" -- 
#> immunarch
#>   -- [12/18] Parsing "/tmp/RtmpVH5iE5/MS6.tsv.gz" -- 
#> immunarch
#>   -- [13/18] Parsing "/tmp/RtmpVH5iE5/file1ed423b690d0" -- 
#> unsupported format, skipping
#>   -- [14/18] Parsing "/tmp/RtmpVH5iE5/file1ed43d0d8dbf" -- 
#> unsupported format, skipping
#>   -- [15/18] Parsing "/tmp/RtmpVH5iE5/file1ed44fefb7c4" -- 
#> unsupported format, skipping
#>   -- [16/18] Parsing "/tmp/RtmpVH5iE5/file1ed471c8510c" -- 
#> unsupported format, skipping
#>   -- [17/18] Parsing "/tmp/RtmpVH5iE5/file1ed4d5cc2fb" -- 
#> unsupported format, skipping
#>   -- [18/18] Parsing "/tmp/RtmpVH5iE5/metadata.txt" -- 
#> metadata
#> Processing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de" ...
#>   -- [1/4] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/bootstrap.bundle.min.js" -- 
#> unsupported format, skipping
#>   -- [2/4] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/bootstrap.bundle.min.js.map" -- 
#> Warning: incomplete final line found on '/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/bootstrap.bundle.min.js.map'
#> unsupported format, skipping
#>   -- [3/4] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/bootstrap.min.css" -- 
#> unsupported format, skipping
#>   -- [4/4] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/font.css" -- 
#> unsupported format, skipping
#> Processing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts" ...
#>   -- [1/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/07d40e985ad7c747025dabb9f22142c4.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [2/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/1Ptug8zYS_SKggPNyC0ITw.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [3/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/1Ptug8zYS_SKggPNyCAIT5lu.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [4/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/1Ptug8zYS_SKggPNyCIIT5lu.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [5/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/1Ptug8zYS_SKggPNyCMIT5lu.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [6/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/1Ptug8zYS_SKggPNyCkIT5lu.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [7/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/1f5e011d6aae0d98fc0518e1a303e99a.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [8/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCs6KVjbNBYlgoKcQ72j00.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [9/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCs6KVjbNBYlgoKcg72j00.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [10/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCs6KVjbNBYlgoKcw72j00.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [11/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCs6KVjbNBYlgoKew72j00.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [12/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCs6KVjbNBYlgoKfA72j00.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [13/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCs6KVjbNBYlgoKfw72.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [14/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCv6KVjbNBYlgoCxCvjs2yNL4U.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [15/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCv6KVjbNBYlgoCxCvjsGyN.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [16/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCv6KVjbNBYlgoCxCvjtGyNL4U.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [17/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCv6KVjbNBYlgoCxCvjvGyNL4U.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [18/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCv6KVjbNBYlgoCxCvjvWyNL4U.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [19/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/4iCv6KVjbNBYlgoCxCvjvmyNL4U.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [20/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/626330658504e338ee86aec8e957426b.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [21/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK1dSBYKcSV-LCoeQqfX1RYOo3qPZ7jsDJT9g.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [22/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK1dSBYKcSV-LCoeQqfX1RYOo3qPZ7ksDJT9g.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [23/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK1dSBYKcSV-LCoeQqfX1RYOo3qPZ7nsDI.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [24/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK1dSBYKcSV-LCoeQqfX1RYOo3qPZ7osDJT9g.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [25/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK1dSBYKcSV-LCoeQqfX1RYOo3qPZ7psDJT9g.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [26/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK1dSBYKcSV-LCoeQqfX1RYOo3qPZ7qsDJT9g.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [27/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK1dSBYKcSV-LCoeQqfX1RYOo3qPZ7rsDJT9g.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [28/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK3dSBYKcSV-LCoeQqfX1RYOo3qN67lqDY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [29/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK3dSBYKcSV-LCoeQqfX1RYOo3qNK7lqDY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [30/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK3dSBYKcSV-LCoeQqfX1RYOo3qNa7lqDY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [31/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK3dSBYKcSV-LCoeQqfX1RYOo3qNq7lqDY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [32/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK3dSBYKcSV-LCoeQqfX1RYOo3qO67lqDY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [33/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK3dSBYKcSV-LCoeQqfX1RYOo3qOK7l.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [34/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xK3dSBYKcSV-LCoeQqfX1RYOo3qPK7lqDY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [35/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3i54rwkxduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [36/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3i54rwlBduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [37/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3i54rwlxdu.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [38/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3i54rwmBduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [39/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3i54rwmRduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [40/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3i54rwmhduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [41/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3i54rwmxduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [42/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ig4vwkxduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [43/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ig4vwlBduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [44/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ig4vwlxdu.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [45/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ig4vwmBduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [46/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ig4vwmRduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [47/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ig4vwmhduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [48/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ig4vwmxduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [49/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ik4zwkxduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [50/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ik4zwlBduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [51/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ik4zwlxdu.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [52/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ik4zwmBduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [53/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ik4zwmRduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [54/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ik4zwmhduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [55/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/6xKydSBYKcSV-LCoeQqfX1RYOo3ik4zwmxduz8A.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [56/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/CSR54z1Qlv-GDxkbKVQ_dFsvWNReuQ.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [57/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/CSR54z1Qlv-GDxkbKVQ_dFsvWNpeudwk.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [58/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/CSR64z1Qlv-GDxkbKVQ_fO4KTet_.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [59/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/CSR64z1Qlv-GDxkbKVQ_fOAKTQ.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [60/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_QiYsKILxRpg3hIP6sJ7fM7PqlONvQlMIXxw.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [61/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_QiYsKILxRpg3hIP6sJ7fM7PqlONvUlMI.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [62/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_QiYsKILxRpg3hIP6sJ7fM7PqlONvXlMIXxw.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [63/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_QiYsKILxRpg3hIP6sJ7fM7PqlONvYlMIXxw.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [64/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_QiYsKILxRpg3hIP6sJ7fM7PqlONvZlMIXxw.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [65/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_QiYsKILxRpg3hIP6sJ7fM7PqlONvalMIXxw.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [66/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_QiYsKILxRpg3hIP6sJ7fM7PqlONvblMIXxw.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [67/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_SiYsKILxRpg3hIP6sJ7fM7PqlM-vWjMY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [68/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_SiYsKILxRpg3hIP6sJ7fM7PqlMOvWjMY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [69/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_SiYsKILxRpg3hIP6sJ7fM7PqlMevWjMY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [70/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_SiYsKILxRpg3hIP6sJ7fM7PqlMuvWjMY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [71/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_SiYsKILxRpg3hIP6sJ7fM7PqlOevWjMY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [72/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_SiYsKILxRpg3hIP6sJ7fM7PqlPevW.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [73/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/HI_SiYsKILxRpg3hIP6sJ7fM7PqlPuvWjMY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [74/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/JTUSjIg1_i6t8kCHKm459W1hyzbi.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [75/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/JTUSjIg1_i6t8kCHKm459WRhyzbi.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [76/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/JTUSjIg1_i6t8kCHKm459WZhyzbi.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [77/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/JTUSjIg1_i6t8kCHKm459Wdhyzbi.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [78/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/JTUSjIg1_i6t8kCHKm459Wlhyw.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [79/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/QGYpz_kZZAGCONcK2A4bGOj8mNhN.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [80/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/S6u8w4BMUTPHjxsAUi-qJCY.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [81/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/S6u8w4BMUTPHjxsAXC-q.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [82/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/S6u9w4BMUTPHh6UVSwaPGR_p.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [83/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/S6u9w4BMUTPHh6UVSwiPGQ.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [84/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/S6u9w4BMUTPHh7USSwaPGR_p.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [85/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/S6u9w4BMUTPHh7USSwiPGQ.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [86/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/S6uyw4BMUTPHjx4wXg.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [87/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/S6uyw4BMUTPHjxAwXjeu.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [88/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/XRXV3I6Li01BKofIMeaBXso.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [89/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/XRXV3I6Li01BKofINeaB.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [90/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/XRXV3I6Li01BKofIO-aBXso.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [91/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/XRXV3I6Li01BKofIOOaBXso.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [92/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/XRXV3I6Li01BKofIOuaBXso.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [93/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/c2f002b3a87d3f9bfeebb23d32cfd9f8.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [94/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/ee91700cdbf7ce16c054c2bb8946c736.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [95/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memtYaGs126MiZpBA-UFUIcVXSCEkx2cmqvXlWqW106F15M.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [96/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memtYaGs126MiZpBA-UFUIcVXSCEkx2cmqvXlWqWt06F15M.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [97/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memtYaGs126MiZpBA-UFUIcVXSCEkx2cmqvXlWqWtE6F15M.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [98/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memtYaGs126MiZpBA-UFUIcVXSCEkx2cmqvXlWqWtU6F15M.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [99/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memtYaGs126MiZpBA-UFUIcVXSCEkx2cmqvXlWqWtk6F15M.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [100/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memtYaGs126MiZpBA-UFUIcVXSCEkx2cmqvXlWqWu06F15M.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [101/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memtYaGs126MiZpBA-UFUIcVXSCEkx2cmqvXlWqWuU6F.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [102/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memtYaGs126MiZpBA-UFUIcVXSCEkx2cmqvXlWqWuk6F15M.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [103/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memtYaGs126MiZpBA-UFUIcVXSCEkx2cmqvXlWqWvU6F15M.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [104/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memtYaGs126MiZpBA-UFUIcVXSCEkx2cmqvXlWqWxU6F15M.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [105/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memvYaGs126MiZpBA-UvWbX2vVnXBbObj2OVTS-muw.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [106/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memvYaGs126MiZpBA-UvWbX2vVnXBbObj2OVTS2mu1aB.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [107/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memvYaGs126MiZpBA-UvWbX2vVnXBbObj2OVTSCmu1aB.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [108/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memvYaGs126MiZpBA-UvWbX2vVnXBbObj2OVTSGmu1aB.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [109/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memvYaGs126MiZpBA-UvWbX2vVnXBbObj2OVTSKmu1aB.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [110/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memvYaGs126MiZpBA-UvWbX2vVnXBbObj2OVTSOmu1aB.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [111/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memvYaGs126MiZpBA-UvWbX2vVnXBbObj2OVTSumu1aB.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [112/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memvYaGs126MiZpBA-UvWbX2vVnXBbObj2OVTSymu1aB.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [113/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memvYaGs126MiZpBA-UvWbX2vVnXBbObj2OVTUGmu1aB.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [114/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/memvYaGs126MiZpBA-UvWbX2vVnXBbObj2OVTVOmu1aB.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [115/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/q5uGsou0JOdh94bfuQltOxU.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#>   -- [116/116] Parsing "/tmp/RtmpVH5iE5/bslib-9b21a61df172a688462472a778d4f1de/fonts/q5uGsou0JOdh94bfvQlt.woff2" -- 
#> Warning: line 1 appears to contain an embedded nul
#> unsupported format, skipping
#> Processing "/tmp/RtmpVH5iE5/downlit" ...
#>   -- [1/10] Parsing "/tmp/RtmpVH5iE5/downlit/base" -- 
#> unsupported format, skipping
#>   -- [2/10] Parsing "/tmp/RtmpVH5iE5/downlit/data.table" -- 
#> unsupported format, skipping
#>   -- [3/10] Parsing "/tmp/RtmpVH5iE5/downlit/dplyr" -- 
#> unsupported format, skipping
#>   -- [4/10] Parsing "/tmp/RtmpVH5iE5/downlit/factoextra" -- 
#> unsupported format, skipping
#>   -- [5/10] Parsing "/tmp/RtmpVH5iE5/downlit/fpc" -- 
#> unsupported format, skipping
#>   -- [6/10] Parsing "/tmp/RtmpVH5iE5/downlit/immundata" -- 
#> unsupported format, skipping
#>   -- [7/10] Parsing "/tmp/RtmpVH5iE5/downlit/pak" -- 
#> unsupported format, skipping
#>   -- [8/10] Parsing "/tmp/RtmpVH5iE5/downlit/stats" -- 
#> unsupported format, skipping
#>   -- [9/10] Parsing "/tmp/RtmpVH5iE5/downlit/tibble" -- 
#> unsupported format, skipping
#>   -- [10/10] Parsing "/tmp/RtmpVH5iE5/downlit/utils" -- 
#> unsupported format, skipping
#> Processing "/tmp/RtmpVH5iE5/duckplyr" ...
#>   -- [1/2] Parsing "/tmp/RtmpVH5iE5/duckplyr/duckplyr1ed45def4ddb.duckdb" -- 
#> Warning: line 1 appears to contain an embedded nul
#> Warning: incomplete final line found on '/tmp/RtmpVH5iE5/duckplyr/duckplyr1ed45def4ddb.duckdb'
#> Error in tolower(l): invalid multibyte string 1
# sum(immdata$data[[1]] != new_immdata$data[[1]], na.rm = TRUE)
# sum(immdata$data[[2]] != new_immdata$data[[2]], na.rm = TRUE)
# sum(immdata$meta != new_immdata$meta, na.rm = TRUE)
# }