Calculates number of mutations against the germline for each clonotype

This function aligns V and J genes from the germline in each cluster with corresponding genes in each clonotype, saves the alignments for purpose of visualization, and calculates number of mutations for each clonotype.

Usage

repSomaticHypermutation(.data, .threads, .nofail)

Arguments

.data

The data to be processed: an output of repClonalFamily(); variants with one sample and list of samples are both supported.

.threads

Number of threads to use.

.nofail

Will return NA instead of stopping if Clustal W is not installed. Used to avoid raising errors in examples on computers where Clustal W is not installed.

Value

Dataframe or list of dataframes (if input is a list with multiple samples). The dataframe has all the columns from repClonalFamily() output dataframe, with Sequence column unnested: the resulting dataframe has one line per clonotype. Clone.ID column contains original IDs for clonotypes, and can be used as dataframe key. New columns are added:

Germline.Alignment.V: contains V gene alignment of current clonotype with the germline

Germline.Alignment.J: contains J gene alignment of current clonotype with the germline

Substitutions: contains number of substitutions in the alignment (summary for V and J)

Insertions: contains number of insertions in the clonotype relative to germline (summary for V and J)

Deletions: contains number of deletions in the clonotype relative to germline (summary for V and J)

Mutations: contains total number of mutations in the alignment (summary for V and J)

Examples

data(bcrdata)
bcr_data <- bcrdata$data

bcr_data %>%
  seqCluster(seqDist(bcr_data), .fixed_threshold = 3) %>%
  repGermline(.threads = 1) %>%
  repAlignLineage(.min_lineage_sequences = 2, .align_threads = 2, .nofail = TRUE) %>%
  repClonalFamily(.threads = 1, .nofail = TRUE) %>%
  repSomaticHypermutation(.threads = 1, .nofail = TRUE)
#> repAlignLineage requires Clustal W app to be installed!
#> Please download it from here: http://www.clustal.org/download/current/
#> or install it with your system package manager (such as apt or dnf).
#> [1] NA
#> attr(,"class")
#> [1] "step_failure_ignored" "logical"