Get the N most abundant clonotypes

Get the N most abundant clonotypes

Usage

top(.data, .n = 10)

Arguments

.data

The data to be processed. Can be data.frame, data.table::data.table, or a list of these objects.

Every object must have columns in the immunarch compatible format. immunarch_data_format

Competent users may provide advanced data representations: DBI database connections, or a list of these objects. They are supported with the same limitations as basic objects.

Note: each connection must represent a separate repertoire.

.n

Numeric. Number of the most abundant clonotypes to return.

Value

Data frame with the .n most abundant clonotypes only.

Examples

data(immdata)
top(immdata$data)
#> $`A2-i129`
#> # A tibble: 10 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1    173    0.0204  TGCGCCAGC… CASSQE… TRBV4… TRBD1  TRBJ2…    16      18    26
#>  2    163    0.0192  TGCGCCAGC… CASSYR… TRBV4… TRBD1  TRBJ2…    11      13    18
#>  3     66    0.00776 TGTGCCACC… CATSTN… TRBV15 TRBD1  TRBJ2…    11      16    22
#>  4     54    0.00635 TGTGCCACC… CATSIG… TRBV15 TRBD2  TRBJ2…    11      19    25
#>  5     48    0.00565 TGTGCCAGC… CASSPW… TRBV27 TRBD1  TRBJ1…    11      16    23
#>  6     48    0.00565 TGCGCCAGC… CASQGD… TRBV4… TRBD1  TRBJ1…     8      13    19
#>  7     40    0.00471 TGCGCCAGC… CASSQD… TRBV4… TRBD1  TRBJ2…    16      21    26
#>  8     31    0.00365 TGTGCCAGC… CASSEE… TRBV2  TRBD1  TRBJ1…    15      17    20
#>  9     30    0.00353 TGCGCCAGC… CASSQP… TRBV4… TRBD1  TRBJ2…    14      23    28
#> 10     28    0.00329 TGTGCCAGC… CASSWV… TRBV6… TRBD1  TRBJ2…    12      20    25
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
#> $`A2-i131`
#> # A tibble: 10 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1    111    0.0131  TGCAGTGCT… CSASRG… TRBV2… TRBD1  TRBJ2…    11      12    17
#>  2     93    0.0109  TGTGCCAGC… CASSVA… TRBV9  TRBD1  TRBJ2…    15      21    23
#>  3     66    0.00776 TGTGCCAGC… CASSRM… TRBV13 TRBD1  TRBJ2…    11      18    24
#>  4     59    0.00694 TGTGCCAGC… CASSPT… TRBV6… TRBD2  TRBJ2…    10      14    19
#>  5     57    0.00671 TGCGCCAGC… CASSLD… TRBV5… TRBD2  TRBJ1…    15      17    20
#>  6     47    0.00553 TGTGCCAGC… CASRGL… TRBV6… TRBD2  TRBJ2…    10      11    16
#>  7     46    0.00541 TGCAGCGTT… CSVTGV… TRBV2… TRBD1  TRBJ2…     8       9    13
#>  8     30    0.00353 TGTGCCAGC… CASSYL… TRBV6… TRBD2  TRBJ1…    15      17    19
#>  9     29    0.00341 TGTGCCAGC… CASSLA… TRBV5… TRBD1  TRBJ1…    15      21    26
#> 10     29    0.00341 TGTGCCAGC… CASSYI… TRBV6… TRBD1  TRBJ1…    14      17    20
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
#> $`A2-i133`
#> # A tibble: 11 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1    333    0.0392  TGTGCCAGC… CASSLY… TRBV1… TRBD2  TRBJ2…    11      18    29
#>  2    111    0.0131  TGTGCCAGC… CASSLG… TRBV7… TRBD1  TRBJ1…    12      15    19
#>  3    102    0.012   TGTGCCAGC… CASSPL… TRBV7… TRBD2  TRBJ2…    11      18    23
#>  4     81    0.00953 TGTGCCAGC… CASSPF… TRBV7… TRBD2  TRBJ1…    11      16    20
#>  5     71    0.00835 TGTGCCAGC… CASSQG… TRBV4… TRBD2  TRBJ2…    15      19    22
#>  6     53    0.00624 TGTGCCAGC… CASSVT… TRBV9  TRBD1  TRBJ1…    13      14    18
#>  7     45    0.00529 TGTGCCAGC… CASSLG… TRBV7… TRBD1  TRBJ1…    15      16    18
#>  8     38    0.00447 TGTGCCAGC… CASSST… TRBV7… TRBD1  TRBJ1…    16      19    24
#>  9     38    0.00447 TGTGCCAGC… CASREL… TRBV28 TRBD2  TRBJ2…    10      24    29
#> 10     21    0.00247 TGCAGCGTT… CSVDEG… TRBV2… TRBD1  TRBJ2…    10      14    17
#> 11     21    0.00247 TGTGCCAGC… CASSLS… TRBV7… TRBD2  TRBJ2…    15      19    21
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
#> $`A2-i132`
#> # A tibble: 10 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1     41    0.00482 TGTGCCAGC… CASSFR… TRBV7… TRBD1  TRBJ2…    10      16    26
#>  2     23    0.00271 TGTGCCATC… CAIRKQ… TRBV1… TRBD1  TRBJ1…    10      15    19
#>  3     20    0.00235 TGCGCCAGC… CASSTR… TRBV4… TRBD2  TRBJ2…    11      20    28
#>  4     20    0.00235 TGTGCCAGC… CASSYL… TRBV6… TRBD1  TRBJ1…    14      18    27
#>  5     18    0.00212 TGTGCCAGC… CASSYI… TRBV6… TRBD1  TRBJ1…    14      18    23
#>  6     18    0.00212 TGTGCCAGC… CASSLL… TRBV28 TRBD2  TRBJ2…    14      15    24
#>  7     17    0.002   TGTGCCAGC… CASTSL… TRBV6… TRBD1  TRBJ2…     9      25    29
#>  8     16    0.00188 TGCGCCAGC… CASSLD… TRBV5… TRBD1  TRBJ2…    15      -1    -1
#>  9     15    0.00176 TGTGCCAGC… CASTLR… TRBV28 TRBD1  TRBJ1…     9      19    21
#> 10     15    0.00176 TGCGCCAGC… CASSQE… TRBV4… TRBD1  TRBJ2…    16      20    23
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
#> $`A4-i191`
#> # A tibble: 10 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1    449    0.0528  TGTGCCAGC… CASSAF… TRBV9  TRBD2  TRBJ2…    12      19    24
#>  2    202    0.0238  TGCGCCTGT… CACGGG… TRBV1… TRBD2  TRBJ1…     5      12    18
#>  3    187    0.022   TGTGCCAGC… CASSSG… TRBV27 TRBD2  TRBJ2…    12      15    23
#>  4    143    0.0168  TGTGCCAGC… CASSSG… TRBV6… TRBD1  TRBJ2…    12      14    21
#>  5    126    0.0148  TGCGCCAGC… CASSQD… TRBV4… TRBD1  TRBJ2…    16      19    22
#>  6     83    0.00976 TGCGCCAGC… CASSQV… TRBV4… TRBD2  TRBJ1…    15      17    21
#>  7     80    0.00941 TGTGCCAGC… CASSLG… TRBV28 TRBD2  TRBJ2…    14      20    24
#>  8     68    0.008   TGTGCCAGC… CASSFG… TRBV1… TRBD1  TRBJ2…    13      23    26
#>  9     66    0.00776 TGTGCCAGC… CASSAG… TRBV7… TRBD1  TRBJ2…    11      15    21
#> 10     56    0.00659 TGCGCCAGC… CASSLA… TRBV5… TRBD2  TRBJ1…    15      20    24
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
#> $`A4-i192`
#> # A tibble: 11 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1    150    0.0176  TGCGCCAGC… CASSES… TRBV1… TRBD1  TRBJ2…    16      17    19
#>  2    124    0.0146  TGTGCCAGC… CASIPP… TRBV7… TRBD2  TRBJ2…     9      30    38
#>  3    120    0.0141  TGCGCCAGC… CASSSG… TRBV4… TRBD1  TRBJ1…    11      15    20
#>  4    115    0.0135  TGCAGTGCT… CSARDI… TRBV2… TRBD1  TRBJ1…    13      18    20
#>  5    107    0.0126  TGTGCCAGC… CASSRW… TRBV7… TRBD1  TRBJ2…    11      16    20
#>  6    101    0.0119  TGCAGTGCT… CSARDF… TRBV2… TRBD1  TRBJ2…    13      24    31
#>  7     95    0.0112  TGTGCCAGC… CASSPN… TRBV1… TRBD2  TRBJ1…    10      11    14
#>  8     76    0.00894 TGTGCCAGC… CASSLI… TRBV27 TRBD1  TRBJ2…    14      17    19
#>  9     51    0.006   TGTGCCAGC… CASSPK… TRBV18 TRBD2  TRBJ2…    14      18    24
#> 10     42    0.00494 TGTGCCAGC… CASSPN… TRBV1… TRBD2  TRBJ1…    11      16    19
#> 11     42    0.00494 TGCGCCAGC… CASSHP… TRBV4… TRBD2  TRBJ2…    13      17    28
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
#> $MS1
#> # A tibble: 10 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1    539    0.0634  TGTGCCAGC… CASSLQ… TRBV7… TRBD2  TRBJ2…    14      18    26
#>  2    320    0.0376  TGTGCCAGC… CASSVY… TRBV9  TRBD1  TRBJ2…    13      20    22
#>  3    204    0.024   TGCAGTACC… CSTEED… TRBV2… TRBD1  TRBJ2…     5      14    19
#>  4    151    0.0178  TGCAGCGTT… CSVELR… TRBV2… TRBD1  TRBJ2…    11      16    19
#>  5     98    0.0115  TGTGCCAGC… CASSLG… TRBV5… TRBD1  TRBJ2…    15      16    22
#>  6     81    0.00953 TGCAGTGCT… CSARDL… TRBV2… TRBD2  TRBJ2…    13      15    20
#>  7     76    0.00894 TGTGCCAGC… CASSFE… TRBV7… TRBD1  TRBJ2…    13      16    22
#>  8     62    0.00729 TGCAGCTAT… CSYRTG… TRBV2… TRBD1  TRBJ2…     5      10    20
#>  9     60    0.00706 TGTGCCAGC… CASSLG… TRBV28 TRBD1  TRBJ2…    14      26    31
#> 10     38    0.00447 TGTGCCAGC… CASSLY… TRBV7… TRBD2  TRBJ2…    14      17    19
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
#> $MS2
#> # A tibble: 10 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1     97    0.0114  TGTGCCAGC… CASSTR… TRBV28 TRBD1  TRBJ2…    10      15    20
#>  2     77    0.00906 TGCAGCGTT… CSVEGG… TRBV2… TRBD1  TRBJ2…    12      13    17
#>  3     33    0.00388 TGTGCCAGC… CASSLI… TRBV28 TRBD1  TRBJ2…    14      20    26
#>  4     32    0.00376 TGTGCCAGC… CASSIT… TRBV27 TRBD1  TRBJ1…    11      16    18
#>  5     26    0.00306 TGCAGCGTT… CSVEGG… TRBV2… TRBD2  TRBJ2…    12      13    16
#>  6     25    0.00294 TGTGCCAGC… CASGSG… TRBV1… TRBD1  TRBJ2…     8      15    23
#>  7     24    0.00282 TGCAGTGCT… CSAMGS… TRBV2… TRBD2  TRBJ2…     9      16    25
#>  8     22    0.00259 TGTGCCAGC… CASSQG… TRBV7… TRBD1  TRBJ2…    11      12    17
#>  9     22    0.00259 TGTGCCAGC… CASSPG… TRBV7… TRBD1  TRBJ2…    11      15    21
#> 10     21    0.00247 TGCAGTGCT… CSAGRD… TRBV2… TRBD1  TRBJ2…     8      13    17
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
#> $MS3
#> # A tibble: 10 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1    608    0.0715  TGTGCCAGC… CASSQD… TRBV14 TRBD1  TRBJ2…    16      19    21
#>  2    291    0.0342  TGTGCCAGC… CASSPT… TRBV1… TRBD1  TRBJ2…    11      14    22
#>  3    131    0.0154  TGTGCCAGC… CASSQD… TRBV14 TRBD2  TRBJ2…    16      18    21
#>  4    115    0.0135  TGTGCCACC… CATSII… TRBV15 TRBD2  TRBJ2…    11      16    24
#>  5     78    0.00918 TGTGCCACC… CATSRL… TRBV15 TRBD2  TRBJ2…    11      13    24
#>  6     51    0.006   TGTGCCAGC… CASSLG… TRBV5… TRBD1  TRBJ2…    15      23    26
#>  7     33    0.00388 TGTGCCAGC… CASSPG… TRBV7… TRBD1  TRBJ2…    11      16    19
#>  8     33    0.00388 TGTGCCAGC… CASSPG… TRBV7… TRBD2  TRBJ2…    11      15    27
#>  9     32    0.00376 TGTGCCATC… CAIRDP… TRBV1… TRBD1  TRBJ2…    10      17    20
#> 10     29    0.00341 TGTGCCAGC… CASSLG… TRBV7… TRBD2  TRBJ2…    15      16    27
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
#> $MS4
#> # A tibble: 10 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1     28    0.00329 TGTGCCAGC… CASSVV… TRBV9  TRBD1  TRBJ2…    15      24    29
#>  2     27    0.00318 TGTGCCAGC… CASSPD… TRBV7… TRBD1  TRBJ2…    11      21    24
#>  3     27    0.00318 TGTGCCACC… CATSRD… TRBV15 TRBD2  TRBJ2…    11      26    31
#>  4     24    0.00282 TGCGCCAGC… CASQRE… TRBV4… TRBD2  TRBJ1…     8      15    18
#>  5     18    0.00212 TGTGCCAGC… CASSHR… TRBV5… TRBD1  TRBJ2…    11      16    25
#>  6     16    0.00188 TGTGCCAGC… CASSFV… TRBV1… TRBD1  TRBJ2…    13      17    23
#>  7     15    0.00176 TGTGCCAGC… CASSYS… TRBV7… TRBD2  TRBJ2…    10      16    20
#>  8     15    0.00176 TGCGCCAGC… CASRSG… TRBV5… TRBD1  TRBJ2…    10      16    26
#>  9     14    0.00165 TGTGCCATC… CAISDR… TRBV1… TRBD1  TRBJ1…    13      14    18
#> 10     14    0.00165 TGCAGCGCC… CSATGD… TRBV2… TRBD1  TRBJ1…     7       9    15
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
#> $MS5
#> # A tibble: 10 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1    433    0.0509  TGTGCCAGC… CASSPG… TRBV1… TRBD1  TRBJ2…    10      13    19
#>  2    414    0.0487  TGTGCCAGC… CASSPT… TRBV6… TRBD1  TRBJ2…    10      16    19
#>  3    272    0.032   TGTGCCAGC… CASSDG… TRBV2… TRBD1  TRBJ1…    13      18    20
#>  4    233    0.0274  TGTGCCAGC… CASSQG… TRBV7… TRBD1  TRBJ1…    11      14    19
#>  5    106    0.0125  TGTGCCAGC… CASSLG… TRBV7… TRBD2  TRBJ2…    15      16    26
#>  6     79    0.00929 TGTGCCAGC… CASRFK… TRBV6… TRBD2  TRBJ2…    10      18    27
#>  7     75    0.00882 TGTGCCAGC… CASSLG… TRBV28 TRBD1  TRBJ2…    14      15    18
#>  8     55    0.00647 TGTGCCAGC… CASSQD… TRBV3… TRBD2  TRBJ2…    16      24    27
#>  9     45    0.00529 TGCGCCAGC… CASSQE… TRBV4… TRBD2  TRBJ1…    16      19    22
#> 10     40    0.00471 TGTGCCACC… CATSRV… TRBV15 TRBD2  TRBJ2…    11      16    23
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
#> $MS6
#> # A tibble: 11 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1     46    0.00541 TGTGCCAGC… CASSAT… TRBV27 TRBD1  TRBJ2…    11      14    18
#>  2     41    0.00482 TGCAGTGCT… CSARDK… TRBV2… TRBD2  TRBJ2…    13      16    25
#>  3     29    0.00341 TGTGCCAGC… CASSPL… TRBV1… TRBD1  TRBJ2…    11      19    23
#>  4     26    0.00306 TGTGCCAGC… CASSPN… TRBV1… TRBD2  TRBJ1…    11      12    14
#>  5     21    0.00247 TGTGCCAGC… CASGTG… TRBV7… TRBD1  TRBJ1…     8       9    16
#>  6     20    0.00235 TGCGCCAGC… CASSLS… TRBV5… TRBD2  TRBJ2…    14      18    27
#>  7     18    0.00212 TGTGCCAGC… CASSPN… TRBV1… TRBD1  TRBJ2…    11      19    21
#>  8     17    0.002   TGTGCCAGC… CASSTG… TRBV6… TRBD2  TRBJ2…    11      14    26
#>  9     12    0.00141 TGTGCCAGC… CASSDS… TRBV6… TRBD2  TRBJ2…    16      17    27
#> 10     11    0.00129 TGCAGTGCT… CSARED… TRBV2… TRBD1  TRBJ1…    13      18    20
#> 11     11    0.00129 TGTGCCAGC… CASSSP… TRBV27 TRBD2  TRBJ2…    12      13    17
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>
#> 
top(immdata$data[[1]])
#> # A tibble: 10 × 15
#>    Clones Proportion CDR3.nt    CDR3.aa V.name D.name J.name V.end D.start D.end
#>     <dbl>      <dbl> <chr>      <chr>   <chr>  <chr>  <chr>  <int>   <int> <int>
#>  1    173    0.0204  TGCGCCAGC… CASSQE… TRBV4… TRBD1  TRBJ2…    16      18    26
#>  2    163    0.0192  TGCGCCAGC… CASSYR… TRBV4… TRBD1  TRBJ2…    11      13    18
#>  3     66    0.00776 TGTGCCACC… CATSTN… TRBV15 TRBD1  TRBJ2…    11      16    22
#>  4     54    0.00635 TGTGCCACC… CATSIG… TRBV15 TRBD2  TRBJ2…    11      19    25
#>  5     48    0.00565 TGTGCCAGC… CASSPW… TRBV27 TRBD1  TRBJ1…    11      16    23
#>  6     48    0.00565 TGCGCCAGC… CASQGD… TRBV4… TRBD1  TRBJ1…     8      13    19
#>  7     40    0.00471 TGCGCCAGC… CASSQD… TRBV4… TRBD1  TRBJ2…    16      21    26
#>  8     31    0.00365 TGTGCCAGC… CASSEE… TRBV2  TRBD1  TRBJ1…    15      17    20
#>  9     30    0.00353 TGCGCCAGC… CASSQP… TRBV4… TRBD1  TRBJ2…    14      23    28
#> 10     28    0.00329 TGTGCCAGC… CASSWV… TRBV6… TRBD1  TRBJ2…    12      20    25
#> # ℹ 5 more variables: J.start <int>, VJ.ins <dbl>, VD.ins <dbl>, DJ.ins <dbl>,
#> #   Sequence <lgl>