One function to visualise them all

Output from every function in immunarch can be visualised with a single function — vis. The vis automatically detects the type of the data and draws a proper visualisation. For example, output from the repOverlap function will be identified as repertoire overlap values and respective visualisation will be chosen without any additional arguments. See "Details" for the list of available visualisations.

Arguments

.data
Pass the output from any immunarch analysis tool to vis().

Details

List of available visualisations for different kinds of data.

Basic analysis:

- Exploratory analysis results (from repExplore) — see vis.immunr_exp_vol;

- Clonality statistics (from repClonality) — see vis.immunr_homeo.

Overlaps and public clonotypes:

- Overlaps (from repOverlap) using heatmaps, circos plots, polar area plots — see vis.immunr_ov_matrix;

- Overlap clustering (from repOverlapAnalysis) — see vis.immunr_hclust;

- Repertoire incremental overlaps (from repOverlap) — see vis.immunr_top_overlap;

- Public repertoire abundance (from pubRep) — vis vis.immunr_public_repertoire.

Gene usage:

- Gene usage statistics (from geneUsage) using bar plots, box plots, treemaps — see vis.immunr_gene_usage;

- Gene usage distances (from geneUsageAnalysis) using heatmaps, circos plots, polar area plots — see vis.immunr_ov_matrix;

- Gene usage clustering (from geneUsageAnalysis) — see vis.immunr_hclust.

Diversity estimation:

- Diversity estimations (from repDiversity) — see vis.immunr_chao1.

Advanced analysis:

- Repertoire dynamics (from trackClonotypes) — see vis.immunr_dynamics;

- Sequence logo plots of amino acid distributions (from kmer_profile) — see vis_logo;

- Kmers distributions (from getKmers) — see vis.immunr_kmers;

- Mutation networks (from mutationNetwork) — see vis.immunr_mutation_network;

- CDR3 amino acid properties, e.g., biophysical (from cdrProp) — see vis.immunr_cdr_prop.

Additionaly, we provide a wrapper functions for visualisations of common data types:

- Any data frames or matrices using heatmaps — see vis_heatmap and vis_heatmap2;

- Any data frames or matrices using circos plots — see vis_circos;

- Any data frames or matrices using polar area plots — see vis_radar.

See also

Examples

# NOT RUN {
# Load the test data
data(immdata)

# Compute and visualise overlaps
ov = repOverlap(immdata$data)
vis(ov)
# }