Histograms and boxplots (general case / gene usage)

Visualise distributions of genes using heatmaps, treemaps or other plots.

# S3 method for immunr_gene_usage
vis(.data, .plot = c("hist", "box", "tree",
  "heatmap", "heatmap2", "circos"), ...)

Arguments

.data
Output from the geneUsage function.

.plot
String specifying the plot type:

- "hist" for histograms using vis_hist;

- "tree" for histograms using vis_treemap;

- "heatmap" for heatmaps using vis_heatmap;

- "heatmap2" for heatmaps using vis_heatmap2;

- "circos" for circos plots using vis_circos.

...
Other arguments passed to corresponding functions depending on the plot type:

- "hist" - passes arguments to vis_hist;

- "box" - passes arguments to vis_box;

- "tree" - passes arguments to vis_treemap and treemap from the treemap package;

- "heatmap" - passes arguments to vis_heatmap;

- "heatmap2" - passes arguments to vis_heatmap2 and heatmap from the "heatmap3" package;

- "circos" - passes arguments to vis_circos and chordDiagram from the "circlize" package.

See also

Examples

# NOT RUN {
data(twb)
gu = geneUsage(twb)

# next two plots are similar
vis(gu, .title = "Gene usage on twins")
vis_hist(gu, .title = "Gene usage on twins")

# next two plots are similar as well
vis(gu, "box", .title = "Gene usage on twins")
vis_box(gu, .title = "Gene usage on twins")
# }