Visualisation of clustering results

Visualisation of clustering results

vis.immunr_hclust(.data, .rect = F, .plot = c("clust", "best"), .return.grob = F)

vis.immunr_kmeans(.data, .point = T, .text = T, .ellipse = T,
.point.size = 2, .text.size = 10, .plot = c("clust", "best"), .return.grob = F)

vis.immunr_dbscan(.data, .point = T, .text = T, .ellipse = T,
.point.size = 2, .text.size = 10, .plot = c("clust", "best"))

Arguments

.data
Clustering results from repOverlapAnalysis or geneUsageAnalysis.

.rect
Passed to fviz_dend - whether to add a rectangle around groups.

.point
If TRUE then plot sample points. Passed to fviz_cluster.

.text
If TRUE then plot text labels. Passed to fviz_cluster.

.ellipse
If TRUE then plot ellipses around all samples. Passed to "ellipse" from fviz_cluster.

.point.size
Size of points, passed to "pointsize" from fviz_cluster.

.text.size
Size of text labels, passed to labelsize from fviz_cluster.

.plot
A character vector of length one or two specifying which plots to visualise. If "clust" than plot only the clustering. If "best" than plot the number of optimal clusters. If both than plot both.

.return.grob
If TRUE then return grob instead of plot.

See also