WIP

WIP

gene_stats()

Value

gene_stats returns all segment gene statistics

Examples

gene_stats()

#>     alias                 species ighd ighj ighv igij igkj igkv iglj iglv traj
#> 1      bt               BosTaurus   21    4   25    0    1    6    5   26   46
#> 2      cd      CamelusDromedarius    0    0    0    0    0    0    0    0    0
#> 3     clf    CanisLupusFamiliaris    0    0    0    0    0    0    0    0    0
#> 4      dr              DanioRerio    7    7    0    3    0    0    0    0    0
#> 5      hs             HomoSapiens   30   13  248    0    5   64    7   69   57
#> 6  macmul           MacacaMulatta   24    7   19    0    4   83    5    0    0
#> 7     mmc    MusMusculusCastaneus    0    0    0    0    0    4    0    0    0
#> 8     mmd   MusMusculusDomesticus    0    0    0    0    0    2    0    0    0
#> 9  musmus             MusMusculus   32    8  225    0    8  109    3    5   42
#> 10     oa OrnithorhynchusAnatinus    3   10    0    0    0    0    0    0    0
#> 11     oc    OryctolagusCuniculus   10   11   39    0    8   26    2   20    0
#> 12     om      OncorhynchusMykiss    9    7    6    0    0    0    0    0    0
#> 13     rn        RattusNorvegicus   30    4  113    0    6  132    2    8    0
#> 14   smth   MusMusculusMolossinus    0    0    0    0    0    1    0    0    0
#> 15   smth     MusMusculusMusculus    0    0    0    0    0    1    0    0    0
#> 16   smth              MusSpretus    0    0    0    0    0    2    0    2    0
#> 17     ss               SusScrofa    5    5   15    0    8   19    4   14    0
#>    trav trbd trbj trbv trdd trdj trdv trgj trgv
#> 1     0    0    0    0    5    3    0    6   15
#> 2     0    0    0    0    0    0    7    2    2
#> 3     0    2    8   19    0    0    0    7    8
#> 4     0    0    0    0    0    0    0    0    0
#> 5    60    3   14   64    3    4    6    4   10
#> 6     0    2   15   58    0    0    0    0    0
#> 7     0    0    0    0    0    0    0    0    0
#> 8     0    0    0    0    0    0    0    0    0
#> 9   145    2   14   23    2    3    7    0   11
#> 10    0    0    0    0    0    0    0    0    0
#> 11    0    0    0    0    0    0    0    0    0
#> 12    0    1    9    0    0    0    0    0    0
#> 13    0    0    0    0    0    0    0    0    0
#> 14    0    0    0    0    0    0    0    0    0
#> 15    0    0    0    0    0    0    0    0    0
#> 16    0    0    0    0    0    0    0    0    0
#> 17    0    0    0    0    0    0    0    0    0

get_genes("hs.trbv", "segment")

#>  [1] "TRBV10-1" "TRBV10-2" "TRBV10-3" "TRBV11-1" "TRBV11-2" "TRBV11-3"
#>  [7] "TRBV12-3" "TRBV12-4" "TRBV12-5" "TRBV13"   "TRBV14"   "TRBV15"  
#> [13] "TRBV16"   "TRBV18"   "TRBV19"   "TRBV2"    "TRBV20-1" "TRBV24-1"
#> [19] "TRBV25-1" "TRBV27"   "TRBV28"   "TRBV29-1" "TRBV3-1"  "TRBV30"  
#> [25] "TRBV4-1"  "TRBV4-2"  "TRBV4-3"  "TRBV5-1"  "TRBV5-4"  "TRBV5-5" 
#> [31] "TRBV5-6"  "TRBV5-8"  "TRBV6-1"  "TRBV6-2"  "TRBV6-3"  "TRBV6-4" 
#> [37] "TRBV6-5"  "TRBV6-6"  "TRBV6-8"  "TRBV6-9"  "TRBV7-2"  "TRBV7-3" 
#> [43] "TRBV7-4"  "TRBV7-6"  "TRBV7-7"  "TRBV7-8"  "TRBV7-9"  "TRBV9"