Dimensionality reduction

Collect a set of principal variables, reducing the number of not important variables to analyse. Dimensionality reduction makes data analysis algorithms work faster and sometimes more accurate, since it also reduces noise in the data. Currently available methods are:

- immunr_pca performs PCA (Principal Component Analysis) using prcomp;

- immunr_mds performs MDS (Multi-Dimensional Scaling) using isoMDS;

- immunr_tsne performs tSNE (t-Distributed Stochastic Neighbour Embedding) using Rtsne.

immunr_pca(.data, .scale = default_scale_fun, .raw = T, .orig = F, .dist = F)

immunr_mds(.data, .scale = default_scale_fun, .raw = T, .orig = F, .dist = T)

immunr_tsne(.data, .perp = 1, .dist = T, ...)

Arguments

.data
A matrix or a data frame with features, distance matrix or output from repOverlapAnalysis or geneUsageAnalysis functions.

.scale
A function to apply to your data before passing it to any of dimensionality reduction algorithms. There is no scaling by default.

.raw
If TRUE then return non-processed output from dimensionality reduction algorithms. Pass FALSE if you want to visualise results.

.orig
If TRUE then return the original result from algorithms. Pass FALSE if you want to visualise results.

.dist
If TRUE then assume ".data" is a distance matrix.

.perp
The perplexity parameter for Rtsne. Sepcifies the number of neighbours each data point must have in the resulting plot.

...
Other parameters passed to Rtsne.

See also