This function aligns all sequences incliding germline that belong to one clonal lineage and one cluster. After clustering, building clonal lineage and germline, the next step is to analyze the degree of mutation and maturity of each clonal lineage. This allows you to find high mature cells and cells with a large number of offspring. The phylogenetic analysis will find mutations that increase the affinity of BCR. Making alignment of the sequence is the first step towards sequence analysis including BCR.

Aligns all sequences incliding germline within each clonal lineage within each cluster

repAlignLineage(.data,
.min_lineage_sequences, .prepare_threads, .align_threads, .verbose_output, .nofail)

Arguments

.data
The data to be processed. Can be data.frame, data.table or a list of these objects.

.min_lineage_sequences
If number of sequences in the same clonal lineage and the same cluster (not including germline) is lower than this threshold, this group of sequences will not be aligned and will not be used in next steps of BCR pipeline (will be saved in output table only if .verbose_output parameter is set to TRUE).

.prepare_threads
Number of threads to prepare results table. High number can cause heavy memory usage!

.align_threads
Number of threads for lineage alignment.

It must have columns in the immunarch compatible format immunarch_data_format, and also must contain 'Cluster' column, which is added by seqCluster() function, and 'Sequence.germline' column, which is added by repGermline() function.

.verbose_output
If TRUE, all output dataframe columns will be included (see documentation about this function return), and unaligned clusters will be included in the output. Setting this to TRUE significantly increases memory usage. If FALSE, only aligned clusters and columns required for repClonalFamily() calculation will be included in the output.

.nofail
Return NA instead of stopping if Clustal W is not installed. Used to avoid raising errors in examples on computers where Clustal W is not installed.

Value

Dataframe or list of dataframes (if input is a list with multiple samples). The dataframe has these columns: * Cluster: cluster name * Germline: germline sequence * Aligned (included if .verbose_output=TRUE): FALSE if this group of sequences was not aligned with lineage (.min_lineage_sequences is below the threshold); TRUE if it was aligned * Alignment: DNAbin object with alignment or DNAbin object with unaligned sequences (if Aligned=FALSE) * V.length (included if .verbose_output=TRUE): shortest length of V gene part outside of CDR3 region in this group of sequences; longer V genes (including germline) are trimmed to this length before alignment * J.length (included if .verbose_output=TRUE): shortest length of J gene part outside of CDR3 region in this group of sequences; longer J genes (including germline) are trimmed to this length before alignment * Sequences (included if .verbose_output=TRUE): nested dataframe containing all sequences for this combination of cluster and germline; it has columns Sequence, V.end, J.start, CDR3.start, CDR3.end; all values taken from the input dataframe

Examples

data(bcrdata)
bcr_data <- bcrdata$data

bcr_data %>%
  seqCluster(seqDist(bcr_data), .fixed_threshold = 3) %>%
  repGermline() %>%
  repAlignLineage(.min_lineage_sequences = 2, .align_threads = 2, .nofail = TRUE)
#> Warning: Alleles IGHV4-55*01 from sample full_clones not found in the reference and will be dropped!
#> Probably, species argument is wrong (current value: HomoSapiens) or the data contains non-BCR genes.
#> $full_clones
#>                              Cluster
#> 1 IGHV4-61/IGHJ4_length_45_cluster_1
#> 2 IGHV4-61/IGHJ4_length_63_cluster_1
#> 3 IGHV5-51/IGHJ4_length_60_cluster_1
#> 4 IGHV5-51/IGHJ5_length_45_cluster_1
#>                                                                                                                                                                                                                                                                                                                                                                                                 Germline
#> 1           CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACTGNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNCTATGGACTACTGGGGTCAAGGAACCTCAGTCACCGTCTCCTCAG
#> 2     CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTANNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNACTACTGGGGTCAAGGAACCTCAGTCACCGTCTCCTCAG
#> 3 GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNCTATGGACTACTGGGGTCAAGGAACCTCAGTCACCGTCTCCTCAG
#> 4                                                GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNACCGTCTCCTCAG
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Alignment
#> 1                                                       -, c, c, -, c, c, c, c, c, c, c, c, c, c, c, a, a, a, g, g, g, g, g, g, g, g, g, a, a, a, a, a, a, g, g, g, g, g, g, g, g, g, a, a, a, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, t, t, t, g, g, g, g, g, g, g, a, g, t, t, t, a, a, a, t, t, t, a, a, g, t, t, t, c, t, c, t, c, c, a, a, a, t, t, t, t, t, c, a, a, a, c, c, c, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, g, g, g, a, a, a, g, a, g, c, c, c, a, t, a, c, c, c, c, c, c, a, t, a, a, t, a, c, c, t, t, t, t, a, a, a, c, c, c, a, a, a, a, a, a, c, c, t, c, c, c, c, c, c, c, c, c, t, t, t, c, c, c, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, a, a, a, g, a, g, a, a, c, g, g, g, t, t, t, c, c, c, g, g, g, a, a, a, g, g, g, t, t, t, c, c, c, a, a, a, c, c, c, c, c, c, a, a, a, t, t, t, a, g, g, t, t, t, c, c, c, a, a, a, g, g, g, t, t, t, a, a, a, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, g, g, g, t, t, g, c, c, c, c, c, c, a, a, a, a, a, g, g, g, g, a, a, a, a, a, a, c, c, g, c, c, c, a, a, a, g, g, g, t, t, t, t, t, t, c, c, g, t, t, t, c, c, c, c, c, c, c, c, c, t, t, t, g, g, g, a, a, a, a, g, a, g, g, a, c, t, t, t, t, t, g, g, g, a, a, a, g, c, c, c, c, c, t, t, t, c, c, c, t, t, t, g, g, g, t, t, t, g, g, g, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, t, t, t, g, g, g, c, c, c, g, g, g, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, g, g, g, g, g, g, c, c, c, c, c, c, g, g, g, t, t, t, g, g, g, t, t, t, a, a, a, t, t, t, t, t, t, a, t, t, c, c, c, t, t, t, g, g, g, n, t, t, n, g, g, n, c, c, n, g, g, n, a, a, n, g, g, n, a, a, n, g, g, n, g, g, n, c, c, n, g, g, n, c, c, n, c, c, n, g, g, n, g, g, n, g, g, n, g, g, n, a, a, n, c, c, n, t, t, n, a, a, n, c, c, n, g, a, n, c, g, n, c, c, n, a, a, n, c, c, n, t, t, n, c, c, n, c, c, n, c, c, n, t, t, n, t, t, n, t, t, n, g, g, n, a, a, n, c, c, n, t, t, n, a, a, n, c, c, n, t, t, n, g, g, n, g, g, n, g, g, n, g, g, c, c, c, t, c, c, a, c, a, t, g, g, g, g, g, g, g, g, a, a, a, c, a, g, t, c, c, a, c, c, c, c, c, t, t, t, g, g, g, g, g, g, g, -, -, g, -, -, t, t, t, c, c, c, a, a, a, a, c, c, g, c, c, g, g, g, a, t, t, a, c, c, -, t, t, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, g, g, g, t, g, g, c, t, t, a, a, a, c, a, a, c, -, -
#> 2 a, -, -, c, -, -, t, -, -, g, -, -, g, g, g, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, t, t, t, g, g, g, g, g, g, g, g, g, t, t, a, a, a, a, t, t, t, a, a, a, t, t, t, c, c, c, t, t, t, a, a, a, t, t, t, t, t, t, a, a, a, c, c, c, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, g, g, g, a, a, a, g, g, c, c, t, t, a, a, a, c, c, c, c, c, c, a, a, a, a, a, a, c, c, c, t, t, t, a, a, a, c, c, c, a, a, a, a, a, a, c, c, c, c, c, c, c, c, c, c, c, c, t, t, t, c, c, c, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, a, a, a, g, g, g, a, a, a, g, g, g, t, t, t, c, c, c, g, g, g, a, a, a, g, g, g, t, t, t, c, c, c, a, a, a, c, g, g, c, c, c, a, a, a, t, t, t, a, a, a, t, t, t, c, c, c, a, a, c, g, t, t, t, t, t, a, a, a, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, g, g, g, t, t, t, c, c, c, c, c, c, a, a, a, a, a, a, g, g, g, a, a, a, a, a, a, c, c, c, c, c, c, a, a, a, g, g, g, t, t, t, t, t, t, c, c, c, t, t, t, c, c, c, c, c, c, c, c, c, t, t, t, g, g, g, a, a, a, a, a, c, g, c, c, c, c, c, t, t, t, g, g, g, a, a, a, g, g, g, c, c, g, t, t, t, c, c, c, t, t, t, g, g, g, t, t, t, g, g, g, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, t, t, t, g, g, g, c, c, c, g, g, g, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, g, g, g, g, g, g, c, c, c, c, c, c, g, g, g, t, t, t, g, g, g, t, t, t, a, a, a, n, t, t, n, t, t, n, t, t, n, t, t, n, t, t, n, g, g, n, t, t, n, g, g, n, c, c, n, g, g, n, a, a, n, g, g, n, a, a, n, g, g, n, a, a, n, t, t, n, c, c, n, a, a, n, t, t, n, c, c, n, a, a, n, c, c, n, a, a, n, g, g, n, t, t, n, t, t, n, t, t, n, c, c, n, t, t, n, a, a, n, t, t, n, g, g, n, a, a, n, t, t, n, a, a, n, c, c, n, t, a, n, a, a, n, g, g, n, t, t, n, g, g, n, g, g, n, t, t, n, t, t, n, a, a, n, t, t, n, t, t, n, a, a, n, t, t, n, c, c, n, c, c, n, c, c, n, t, t, n, a, a, n, c, c, n, t, t, n, t, t, n, t, t, n, g, g, n, a, a, n, c, c, n, t, t, n, a, a, a, -, -, c, -, -, t, -, -, a, -, -, c, c, c, t, t, t, g, g, g, g, g, g, g, g, g, g, g, g, t, c, c, c, c, c, a, a, a, a, g, g, g, g, g, g, g, g, a, a, a, a, a, a, c, c, c, c, c, c, t, c, c, c, t, t, a, g, g, g, g, g, t, t, t, c, c, c, a, a, a, c, c, c, c, c, c, g, g, g, t, t, t, c, c, c, t, t, t, c, c, c, c, c, c, t, t, t, -, c, c, -, a, a, -, g, g, -, g, g, -, t, t, -, a, a, -, a, a, -, g, g
#> 3                                                                g, g, g, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, g, g, g, g, g, g, g, g, g, g, g, a, a, a, a, t, t, t, c, c, c, a, a, a, t, t, t, c, c, c, t, t, t, a, a, a, t, t, t, c, c, c, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, g, g, g, a, a, a, c, c, c, t, t, t, c, c, c, t, t, t, g, g, g, a, a, a, t, t, t, a, a, a, c, c, c, c, c, c, a, a, a, g, g, g, a, a, a, t, t, t, a, a, a, c, c, c, a, a, a, g, g, g, c, c, c, c, c, c, c, c, c, g, g, g, t, t, t, c, c, c, c, c, c, t, t, t, t, t, t, c, c, c, c, c, c, a, a, a, a, a, a, g, g, g, g, a, a, c, c, c, c, c, c, a, a, a, g, g, g, g, g, g, t, t, t, c, c, c, a, a, a, c, c, g, c, c, c, a, a, a, t, t, t, c, g, g, t, t, t, c, c, c, a, a, a, g, g, g, c, t, t, c, c, c, g, g, g, a, g, g, c, c, c, a, a, a, a, a, a, g, g, g, t, t, t, c, c, c, c, c, c, a, a, a, t, t, t, c, c, c, a, g, g, g, g, g, c, c, c, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, c, c, c, t, t, t, a, a, a, c, c, c, c, c, c, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, c, c, c, a, a, a, g, g, g, c, c, c, c, c, c, t, t, t, g, g, g, a, a, a, a, a, a, g, g, g, g, g, g, c, c, c, c, c, c, t, t, t, c, c, c, g, g, g, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, c, c, c, a, a, a, t, t, t, g, g, g, t, t, t, a, a, a, t, t, t, t, t, t, a, a, a, c, c, c, n, t, t, n, g, g, n, t, t, n, g, g, n, t, t, n, g, a, n, a, a, n, g, g, n, a, a, n, c, c, n, a, a, n, c, c, n, t, t, n, c, c, n, t, t, n, g, g, n, g, g, n, t, t, n, a, a, n, a, a, n, g, g, n, a, a, n, t, t, n, a, a, n, g, g, n, a, a, n, a, a, n, t, t, n, c, c, n, a, a, n, t, t, n, c, c, n, t, t, n, g, g, n, t, t, n, g, g, n, g, g, n, g, g, n, c, c, n, c, c, n, c, c, n, c, c, n, a, a, n, g, g, n, a, a, n, a, a, n, g, c, n, c, c, n, t, t, n, t, t, n, t, t, n, g, g, n, a, a, n, c, c, n, t, t, n, a, a, n, c, c, n, t, t, n, g, g, -, g, g, -, g, g, n, g, g, c, c, c, t, c, c, a, a, a, t, g, g, g, g, g, g, g, g, a, a, a, c, a, a, t, c, c, a, c, c, c, c, c, t, t, t, g, g, g, g, g, g, g, -, -, g, -, -, t, t, t, c, c, c, a, a, a, a, c, c, g, c, c, g, g, g, a, t, t, a, c, c, -, t, t, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, g, g, g, t, g, g, c, t, t, a, a, a, c, -, -
#> 4                                                                                                                                                                                                                                                                      c, c, c, c, c, c, c, c, c, g, g, g, g, g, g, g, g, a, a, a, a, a, a, a, a, a, a, g, g, g, g, g, g, c, c, c, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, t, g, t, t, t, g, g, g, g, g, g, g, g, g, g, g, a, a, a, t, t, g, c, c, c, g, a, a, a, t, t, t, c, c, c, t, t, t, a, a, a, t, c, t, c, c, c, c, c, c, t, t, t, g, c, c, g, g, g, t, t, t, g, g, g, a, a, a, c, c, c, t, t, t, c, c, c, t, t, t, g, g, g, a, a, a, t, c, t, a, a, a, c, c, t, c, t, c, a, a, a, g, g, g, a, a, a, t, t, t, a, a, a, c, c, c, a, a, a, g, g, a, c, t, t, c, c, c, c, c, c, g, g, g, t, t, t, c, c, c, c, c, c, t, t, t, t, t, t, c, c, c, c, c, c, a, a, a, a, a, a, g, g, g, g, g, g, c, c, c, c, c, c, a, a, a, g, g, a, g, g, g, t, t, t, c, c, c, a, a, a, c, c, c, c, c, c, a, a, a, t, t, t, c, c, c, t, t, t, c, c, c, a, a, a, g, g, g, c, c, c, c, c, c, g, g, g, a, a, a, c, c, c, a, a, a, a, a, g, g, g, g, t, t, t, c, c, c, c, c, c, a, c, a, t, t, c, c, c, c, a, a, a, g, g, c, c, t, c, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, c, c, c, t, t, t, a, a, a, c, c, c, c, t, c, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, a, a, g, a, g, c, c, c, a, a, a, g, g, g, c, c, t, c, c, c, t, t, t, g, g, g, a, c, a, a, a, a, g, g, g, g, g, g, c, c, c, c, c, c, t, t, t, c, c, c, g, g, g, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, c, c, c, a, a, a, t, t, t, g, c, a, t, t, t, a, a, a, t, c, t, t, t, t, a, a, a, c, c, c, n, t, t, n, g, g, n, t, t, n, g, a, n, c, c, n, g, g, n, a, a, n, g, g, n, a, a, n, c, c, n, a, a, n, a, a, n, c, c, n, a, a, n, c, c, n, c, c, n, c, c, n, t, t, n, t, t, n, t, t, n, a, a, n, g, g, n, a, a, n, t, t, n, a, a, n, g, g, n, t, t, n, c, c, n, g, g, n, t, t, n, g, g, n, t, t, n, c, c, n, t, t, n, g, g, n, g, g, n, t, t, n, a, a, n, c, c, n, a, a, n, c, c, n, t, t, n, t, t, n, g, g, n, g, g, a, g, g, c, g, g, c, c, c, -, c, c, -, a, a, g, g, g, t, g, g, c, g, g, t, a, a, c, a, g, c, c, c, t, c, c, c, c, c, a, -, -, g, -, -
#>