Aligns all sequences incliding germline within each clonal lineage within each cluster

This function aligns all sequences (incliding germline) that belong to one clonal lineage and one cluster. After clustering and building the clonal lineage and germline, the next step is to analyze the degree of mutation and maturity of each clonal lineage. This allows for finding high mature cells and cells with a large number of offspring. The phylogenetic analysis will find mutations that increase the affinity of BCR. Making alignment of the sequence is the first step towards sequence analysis including BCR.

repAlignLineage(.data, .min_lineage_sequences, .prepare_threads, .align_threads, .nofail)

Arguments

.data

The data to be processed. Can be data.frame, data.table or a list of these objects.

.min_lineage_sequences

If number of sequences in the same clonal lineage and the same cluster (not including germline) is lower than this threshold, this group of sequences will be filtered out from the dataframe; so only large enough lineages will be included.

.prepare_threads

Number of threads to prepare results table. Please note that high number can cause heavy memory usage!

.align_threads

Number of threads for lineage alignment.

It must have columns in the immunarch compatible format immunarch_data_format, and also must contain 'Cluster' column, which is added by seqCluster() function, and 'Germline.sequence' column, which is added by repGermline() function.

.nofail

Will return NA instead of stopping if Clustal W is not installed. Used to avoid raising errors in examples on computers where Clustal W is not installed.

Value

Dataframe or list of dataframes (if input is a list with multiple samples). The dataframe has these columns: * Cluster: cluster name * Germline: germline sequence * Alignment: DNAbin object with alignment * Sequences: nested dataframe containing all sequences for this combination of cluster and germline; it has columns * Sequence, CDR1.nt, CDR2.nt, CDR3.nt, FR1.nt, FR2.nt, FR3.nt, FR4.nt, V.allele, J.allele, V.aa, J.aa: all values taken from the input dataframe * Clone.ID: taken from the input dataframe, or created (filled with row numbers) if missing * Clones: taken from the input dataframe, or created (filled with '1' values) if missing

Examples

data(bcrdata)
bcr_data <- bcrdata$data

bcr_data %>%
  seqCluster(seqDist(bcr_data), .fixed_threshold = 3) %>%
  repGermline(.threads = 1) %>%
  repAlignLineage(.min_lineage_sequences = 2, .align_threads = 2, .nofail = TRUE)
#> Warning: Genes or alleles IGHV4-55 from sample full_clones not found in the reference and will be dropped!
#> Probably, species argument is wrong (current value: HomoSapiens) or the data contains non-BCR genes.
#> $full_clones
#>                              Cluster
#> 1 IGHV4-61/IGHJ4_length_45_cluster_1
#> 2 IGHV4-61/IGHJ4_length_63_cluster_1
#> 3 IGHV5-51/IGHJ4_length_60_cluster_1
#> 4 IGHV5-51/IGHJ5_length_45_cluster_1
#>                                                                                                                                                                                                                                                                                                                                                                                         Germline
#> 1                   CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 2 CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 3       GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 4                      GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Alignment
#> 1                                                                                                                                                                   c, c, c, a, a, a, g, g, g, g, g, g, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, c, c, c, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, g, g, g, a, a, a, g, g, g, t, t, t, c, c, c, g, g, g, g, g, g, g, g, g, c, c, c, c, c, c, c, c, c, a, a, a, g, g, g, g, g, g, a, a, a, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, g, g, g, a, a, a, a, a, a, g, g, g, c, c, c, c, c, c, t, t, t, t, t, t, c, c, c, g, g, g, g, g, g, a, a, a, g, g, g, a, a, a, c, c, c, c, c, c, c, c, c, t, t, t, g, g, g, t, t, t, c, c, c, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, c, c, c, c, c, c, t, t, t, g, g, g, c, c, c, a, a, a, c, c, c, t, t, t, g, g, g, t, t, t, c, c, c, t, t, t, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, g, g, g, g, g, g, c, c, c, t, t, t, c, c, c, c, c, c, g, g, g, t, t, t, c, c, c, a, a, a, g, g, g, c, c, c, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, t, t, t, a, g, g, g, g, g, t, t, t, t, t, t, a, a, a, c, c, c, t, t, t, a, a, a, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, c, c, c, c, c, c, g, g, g, g, g, g, c, c, c, a, a, a, g, g, g, c, c, c, c, c, c, c, c, c, c, c, c, c, c, c, a, a, a, g, g, g, g, g, g, g, g, g, a, a, a, a, a, a, g, g, g, g, g, g, g, g, g, a, a, a, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, t, t, t, g, g, g, g, g, g, g, g, a, t, t, t, a, a, a, t, t, t, a, g, a, t, t, t, c, c, t, t, c, c, a, a, a, t, t, t, t, c, t, a, a, a, c, c, c, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, g, g, g, a, a, a, g, g, a, c, c, c, a, a, t, c, c, c, c, c, c, a, a, t, a, a, t, c, t, c, t, t, t, a, a, a, c, c, c, a, a, a, a, a, a, c, t, c, c, c, c, c, c, c, c, c, c, t, t, t, c, c, c, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, a, a, a, g, g, a, a, c, a, g, g, g, t, t, t, c, c, c, g, g, g, a, a, a, g, g, g, t, t, t, c, c, c, a, a, a, c, c, c, c, c, c, a, a, a, t, t, t, a, g, g, t, t, t, c, c, c, a, a, a, g, g, g, t, t, t, a, a, a, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, g, g, g, t, g, t, c, c, c, c, c, c, a, a, a, a, g, a, g, g, g, a, a, a, a, a, a, c, g, c, c, c, c, a, a, a, g, g, g, t, t, t, t, t, t, c, g, c, t, t, t, c, c, c, c, c, c, c, c, c, t, t, t, g, g, g, a, a, a, a, a, g, g, a, g, c, t, t, t, t, t, g, g, g, a, a, a, g, c, c, c, c, c, t, t, t, c, c, c, t, t, t, g, g, g, t, t, t, g, g, g, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, t, t, t, g, g, g, c, c, c, g, g, g, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, g, g, g, g, g, g, c, c, c, c, c, c, g, g, g, t, t, t, g, g, g, t, t, t, a, a, a, t, t, t, t, t, t, a, t, t, c, c, c, n, t, t, n, g, g, n, t, t, n, g, g, n, c, c, n, g, g, n, a, a, n, g, g, n, a, a, n, g, g, n, g, g, n, c, c, n, g, g, n, c, c, n, c, c, n, g, g, n, g, g, n, g, g, n, g, g, n, a, a, n, c, c, n, t, t, n, a, a, n, c, c, n, a, g, n, g, c, n, c, c, n, a, a, n, c, c, n, t, t, n, c, c, n, c, c, n, c, c, n, t, t, n, t, t, n, t, t, n, g, g, n, a, a, n, c, c, n, t, t, n, a, a, n, c, c, n, t, t, n, g, g, n, g, g, g, g, g, g, g, g, c, c, c, c, c, c, a, a, c, a, g, g, g, g, g, g, g, g, a, a, a, a, g, a, c, c, c, c, c, c, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, c, c, c, a, a, a, c, c, c, c, c, c, g, g, g, t, t, t, c, c, c, t, t, t, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, g, g, g
#> 2 c, c, c, a, a, a, g, g, g, g, g, g, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, c, c, c, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, g, g, g, a, a, a, g, g, g, t, t, t, c, c, c, g, g, g, g, g, g, g, g, g, c, c, c, c, c, c, c, c, c, a, a, a, g, g, g, g, g, g, a, a, a, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, g, g, g, a, a, a, a, a, a, g, g, g, c, c, c, c, c, c, t, t, t, t, t, t, c, c, c, g, g, g, g, g, g, a, a, a, g, g, g, a, a, a, c, c, c, c, c, c, c, c, c, t, t, t, g, g, g, t, t, t, c, c, c, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, c, c, c, c, c, c, t, t, t, g, g, g, c, c, c, a, a, a, c, c, c, t, t, t, g, g, g, t, t, t, c, c, c, t, t, t, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, g, g, g, g, g, g, c, c, c, t, t, t, c, c, c, c, c, c, g, g, g, t, t, t, c, c, c, a, a, a, g, g, g, c, c, c, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, t, t, t, a, g, g, g, g, g, t, t, t, t, t, t, a, a, a, c, c, c, t, t, t, a, a, a, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, c, c, c, c, c, c, g, g, g, g, g, g, c, c, c, a, a, a, g, g, g, c, c, c, c, c, c, c, c, c, c, c, c, c, c, c, a, a, a, g, g, g, g, g, g, g, g, g, a, a, a, a, a, a, g, g, g, g, g, g, g, g, g, a, a, a, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, t, t, t, g, g, g, g, g, g, g, g, g, t, t, a, a, a, a, t, t, t, a, a, a, t, t, t, c, c, c, t, t, t, a, a, a, t, t, t, t, t, t, a, a, a, c, c, c, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, g, g, g, a, a, a, g, g, c, c, t, t, a, a, a, c, c, c, c, c, c, a, a, a, a, a, a, c, c, c, t, t, t, a, a, a, c, c, c, a, a, a, a, a, a, c, c, c, c, c, c, c, c, c, c, c, c, t, t, t, c, c, c, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, a, a, a, g, g, g, a, a, a, g, g, g, t, t, t, c, c, c, g, g, g, a, a, a, g, g, g, t, t, t, c, c, c, a, a, a, c, g, g, c, c, c, a, a, a, t, t, t, a, a, a, t, t, t, c, c, c, a, a, c, g, t, t, t, t, t, a, a, a, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, g, g, g, t, t, t, c, c, c, c, c, c, a, a, a, a, a, a, g, g, g, a, a, a, a, a, a, c, c, c, c, c, c, a, a, a, g, g, g, t, t, t, t, t, t, c, c, c, t, t, t, c, c, c, c, c, c, c, c, c, t, t, t, g, g, g, a, a, a, a, a, c, g, c, c, c, c, c, t, t, t, g, g, g, a, a, a, g, g, g, c, c, g, t, t, t, c, c, c, t, t, t, g, g, g, t, t, t, g, g, g, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, t, t, t, g, g, g, c, c, c, g, g, g, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, g, g, g, g, g, g, c, c, c, c, c, c, g, g, g, t, t, t, g, g, g, t, t, t, a, a, a, t, t, t, t, t, t, a, t, t, c, t, t, n, t, t, n, g, g, n, t, t, n, g, g, n, c, c, n, g, g, n, a, a, n, g, g, n, a, a, n, g, g, n, a, a, n, t, t, n, c, c, n, a, a, n, t, t, n, c, c, n, a, a, n, c, c, n, a, a, n, g, g, n, t, t, n, t, t, n, t, t, n, c, c, n, t, t, n, a, a, n, t, t, n, g, g, n, a, a, n, t, t, n, a, a, n, c, c, n, t, a, n, a, a, n, g, g, n, t, t, n, g, g, n, g, g, n, t, t, n, t, t, n, a, a, n, t, t, n, t, t, n, a, a, n, t, t, n, c, c, n, c, c, n, c, c, n, t, t, n, a, a, n, c, c, n, t, t, n, t, t, n, t, t, n, g, g, n, a, a, n, c, c, n, t, t, n, a, a, n, c, c, n, t, t, n, g, g, n, g, g, g, g, g, g, g, g, c, c, c, c, c, c, a, a, a, a, g, g, g, g, g, g, g, g, a, a, a, a, a, a, c, c, c, c, c, c, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, c, c, c, a, a, a, c, c, c, c, c, c, g, g, g, t, t, t, c, c, c, t, t, t, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, g, g, g
#> 3                                                       g, g, g, a, a, a, g, g, g, g, g, g, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, t, t, t, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, c, c, c, a, a, a, g, g, g, a, a, a, g, g, g, g, g, g, t, t, t, g, g, g, a, a, a, a, a, a, a, a, a, a, a, a, a, a, a, g, g, g, c, c, c, c, c, c, c, c, c, g, g, g, g, g, g, g, g, g, g, g, g, a, a, a, g, g, g, t, t, t, c, c, c, t, t, t, c, c, c, t, t, t, g, g, g, a, a, a, a, a, a, g, g, g, a, a, a, t, t, t, c, c, c, t, t, t, c, c, c, c, c, c, t, t, t, g, g, g, t, t, t, a, a, a, a, a, a, g, g, g, g, g, g, g, g, g, t, t, t, t, t, t, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, a, a, a, c, c, c, a, a, a, g, g, g, c, c, c, t, t, t, t, t, t, t, t, t, a, a, a, c, c, c, c, c, c, a, a, a, g, g, g, c, c, c, t, t, t, a, a, a, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, c, c, c, g, g, g, g, g, g, c, c, c, t, t, t, g, g, g, g, g, g, g, g, g, t, t, t, g, g, g, c, c, c, g, g, g, c, c, c, c, c, c, a, a, a, g, g, g, a, a, a, t, t, t, g, g, g, c, c, c, c, c, c, c, c, c, g, g, g, g, g, g, g, g, g, a, a, a, a, a, a, a, a, a, g, g, g, g, g, g, c, c, c, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, g, g, g, g, g, g, g, g, g, g, g, a, a, a, a, t, t, t, c, c, c, a, a, a, t, t, t, c, c, c, t, t, t, a, a, a, t, t, t, c, c, c, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, g, g, g, a, a, a, c, c, c, t, t, t, c, c, c, t, t, t, g, g, g, a, a, a, t, t, t, a, a, a, c, c, c, c, c, c, a, a, a, g, g, g, a, a, a, t, t, t, a, a, a, c, c, c, a, a, a, g, g, g, c, c, c, c, c, c, c, c, c, g, g, g, t, t, t, c, c, c, c, c, c, t, t, t, t, t, t, c, c, c, c, c, c, a, a, a, a, a, a, g, g, g, g, a, a, c, c, c, c, c, c, a, a, a, g, g, g, g, g, g, t, t, t, c, c, c, a, a, a, c, c, g, c, c, c, a, a, a, t, t, t, c, g, g, t, t, t, c, c, c, a, a, a, g, g, g, c, t, t, c, c, c, g, g, g, a, g, g, c, c, c, a, a, a, a, a, a, g, g, g, t, t, t, c, c, c, c, c, c, a, a, a, t, t, t, c, c, c, a, g, g, g, g, g, c, c, c, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, c, c, c, t, t, t, a, a, a, c, c, c, c, c, c, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, c, c, c, a, a, a, g, g, g, c, c, c, c, c, c, t, t, t, g, g, g, a, a, a, a, a, a, g, g, g, g, g, g, c, c, c, c, c, c, t, t, t, c, c, c, g, g, g, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, c, c, c, a, a, a, t, t, t, g, g, g, t, t, t, a, a, a, t, t, t, t, t, t, a, a, a, c, c, c, n, t, t, n, g, g, n, t, t, n, g, g, n, t, t, n, g, a, n, a, a, n, g, g, n, a, a, n, c, c, n, a, a, n, c, c, n, t, t, n, c, c, n, t, t, n, g, g, n, g, g, n, t, t, n, a, a, n, a, a, n, g, g, n, a, a, n, t, t, n, a, a, n, g, g, n, a, a, n, a, a, n, t, t, n, c, c, n, a, a, n, t, t, n, c, c, n, t, t, n, g, g, n, t, t, n, g, g, n, g, g, n, g, g, n, c, c, n, c, c, n, c, c, n, c, c, n, a, a, n, g, g, n, a, a, n, a, a, n, g, c, n, c, c, n, t, t, n, t, t, n, t, t, n, g, g, n, a, a, n, c, c, n, t, t, n, a, a, n, c, c, n, t, t, n, g, g, n, g, g, g, g, g, g, g, g, c, c, c, c, c, c, a, a, a, a, g, g, g, g, g, g, g, g, a, a, a, a, a, a, c, c, c, c, c, c, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, c, c, c, a, a, a, c, c, c, c, c, c, g, g, g, t, t, t, c, c, c, t, t, t, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, g, g, g
#> 4                                                                                                                                                                                              g, g, g, a, a, a, g, g, g, g, g, g, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, t, t, t, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, c, c, c, a, a, a, g, g, g, a, a, a, g, g, g, g, g, g, t, t, t, g, g, g, a, a, a, a, a, a, a, a, a, a, a, a, a, a, a, g, g, g, c, c, c, c, c, c, c, c, c, g, g, g, g, g, g, g, g, g, g, g, g, a, a, a, g, g, g, t, t, t, c, c, c, t, t, t, c, c, c, t, t, t, g, g, g, a, a, a, a, a, a, g, g, g, a, a, a, t, t, t, c, c, c, t, t, t, c, c, c, c, c, c, t, t, t, g, g, g, t, t, t, a, a, a, a, a, a, g, g, g, g, g, g, g, g, g, t, t, t, t, t, t, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, a, a, a, c, c, c, a, a, a, g, g, g, c, c, c, t, t, t, t, t, t, t, t, t, a, a, a, c, c, c, c, c, c, a, a, a, g, g, g, c, c, c, t, t, t, a, a, a, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, t, t, t, c, c, c, g, g, g, g, g, g, c, c, c, t, t, t, g, g, g, g, g, g, g, g, g, t, t, t, g, g, g, c, c, c, g, g, g, c, c, c, c, c, c, a, a, a, g, g, g, a, a, a, t, t, t, g, g, g, c, c, c, c, c, c, c, c, c, g, g, g, g, g, g, g, a, g, a, a, a, a, a, a, a, a, a, g, g, g, g, g, g, c, c, c, c, c, c, t, t, t, g, g, g, g, g, g, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, g, t, t, t, t, g, g, g, g, g, g, g, g, g, g, a, g, a, t, a, t, c, g, c, g, c, a, a, a, t, t, t, c, c, c, t, t, t, a, a, a, t, t, c, c, c, c, c, c, c, t, t, t, g, c, c, g, g, g, t, t, t, g, g, g, a, a, a, c, c, c, t, t, t, c, c, c, t, t, t, g, g, g, a, a, a, t, t, c, a, a, a, c, t, c, c, c, t, a, a, a, g, g, g, a, a, a, t, t, t, a, a, a, c, c, c, a, a, a, g, a, g, c, t, t, c, c, c, c, c, c, g, g, g, t, t, t, c, c, c, c, c, c, t, t, t, t, t, t, c, c, c, c, c, c, a, a, a, a, a, a, g, g, g, g, g, g, c, c, c, c, c, c, a, a, a, g, a, g, g, g, g, t, t, t, c, c, c, a, a, a, c, c, c, c, c, c, a, a, a, t, t, t, c, c, c, t, t, t, c, c, c, a, a, a, g, g, g, c, c, c, c, c, c, g, g, g, a, a, a, c, c, c, a, a, a, a, g, a, g, g, g, t, t, t, c, c, c, c, c, c, a, a, c, t, c, t, c, c, c, a, a, a, g, c, g, c, c, t, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, c, c, c, t, t, t, a, a, a, c, c, c, c, c, t, t, t, t, g, g, g, c, c, c, a, a, a, g, g, g, t, t, t, g, g, g, g, g, g, a, a, a, g, g, a, c, c, c, a, a, a, g, g, g, c, t, c, c, c, c, t, t, t, g, g, g, a, a, c, a, a, a, g, g, g, g, g, g, c, c, c, c, c, c, t, t, t, c, c, c, g, g, g, g, g, g, a, a, a, c, c, c, a, a, a, c, c, c, c, c, c, g, g, g, c, c, c, c, c, c, a, a, a, t, t, t, g, a, c, t, t, t, a, a, a, t, t, c, t, t, t, a, a, a, c, c, c, n, t, t, n, g, g, n, t, t, n, a, g, n, c, c, n, g, g, n, a, a, n, g, g, n, a, a, n, c, c, n, a, a, n, a, a, n, c, c, n, a, a, n, c, c, n, c, c, n, c, c, n, t, t, n, t, t, n, t, t, n, a, a, n, g, g, n, a, a, n, t, t, n, a, a, n, g, g, n, t, t, n, c, c, n, g, g, n, t, t, n, g, g, n, t, t, n, c, c, n, t, t, n, g, g, n, g, g, n, t, t, n, a, a, n, c, c, n, a, a, n, c, c, n, t, t, n, t, t, n, g, g, n, g, g, g, g, g, g, g, g, c, c, c, c, c, c, a, a, a, a, g, g, g, g, g, g, g, g, a, a, a, a, g, a, c, c, c, c, c, c, c, c, c, t, t, t, g, g, g, g, g, g, t, t, t, c, c, c, a, a, a, c, c, c, c, c, c, g, g, g, t, t, t, c, c, c, t, t, t, c, c, c, c, c, c, t, t, t, c, c, c, a, a, a, g, g, g
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Sequences
#> 1                                                                          CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTGGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATGTCCATCACAGTGGGAGCACCAATTACAATCCCTCCCTCAAGCGTCGAGTCACCATGTCAGTAGACACGGCCAGGAAGCAGTTGTCCCTGAAATTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACAGCACTCCCTTTGACTACTGGGGCCAGGGAGCCCTGGTCACCGTCTCCTCAG, CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTGGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGATATATTCATTACAGTGGGAACTCCTTCTACAACCCCTCCCTCAAAAGTCGAGTCACCATGTCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAGGTTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACGCCACTCCCTTTGACTACTGGGGCCCGGGAACCCTGGTCACCGTCTCCTCAG, 42, 643, 26, 301, IGHV4-61*01, IGHV4-61*01, IGHJ4*01, IGHJ4*01, ggtggctccgtcagcagtggtggttactac, ggtggctccgtcagcagtggtggttactac, GTCCATCACAGTGGGAGCACC, ATTCATTACAGTGGGAACTCC, TGTGCGAGAGGCGCCGGGGACTACAGCACTCCCTTTGACTACTGG, TGTGCGAGAGGCGCCGGGGACTACGCCACTCCCTTTGACTACTGG, caggtgcagctgcaggagtcgggcccaggactggtgaagccttcggagaccctgtccctcacctgcactgtctct, caggtgcagctgcaggagtcgggcccaggactggtgaagccttcggagaccctgtccctcacctgcactgtctct, tggagctggatccggcagCCCCCAGGGAAGGGACTGGAGTGGATTGGGTAT, tggagctggatccggcagCCCCCAGGGAAGGGACTGGAGTGGATTGGATAT, AATTACAATCCCTCCCTCAAGCGTCGAGTCACCATGTCAGTAGACACGGCCAGGAAGCAGTTGTCCCTGAAATTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTC, TTCTACAACCCCTCCCTCAAAAGTCGAGTCACCATGTCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAGGTTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTC, GGCCAGGGAGCCCTGGTCACCGTCTCCTCAG, GGCCCGGGAACCCTGGTCACCGTCTCCTCAG, QVQLQESGPGLVKPSETLSLTCTVSGGSVSSGGYYWSWIRQPPGKGLEWIGYVHHSGSTNYNPSLKRRVTMSVDTARKQLSLKLTSVTAADTAVYF, QVQLQESGPGLVKPSETLSLTCTVSGGSVSSGGYYWSWIRQPPGKGLEWIGYIHYSGNSFYNPSLKSRVTMSVDTSKNQFSLRLTSVTAADTAVYF, GQGALV~SPSPQ, GPGTLV~SPSPQ
#> 2 CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTGGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCATTAGACACGTCCAAGAACCAGTTCTCCCTGAACCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACTAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTGGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGAATATCTATTACAGTGGGACTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCCTTAGACACGTCCAAGAACCAGTTCTCCCTGACCCTGAGGTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACAAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, 119, 123, 290, 53, IGHV4-61*01, IGHV4-61*01, IGHJ4*01, IGHJ4*01, ggtggctccgtcagcagtggtggttactac, ggtggctccgtcagcagtggtggttactac, ATCTATTACAGTGGGAGTACC, ATCTATTACAGTGGGACTACC, TGTGCGAGAGATCATCACAGTTTCTATGATACTAGTGGTTATTATCCCTACTTTGACTACTGG, TGTGCGAGAGATCATCACAGTTTCTATGATACAAGTGGTTATTATCCCTACTTTGACTACTGG, caggtgcagctgcaggagtcgggcccaggactggtgaagccttcggagaccctgtccctcacctgcactgtctct, caggtgcagctgcaggagtcgggcccaggactggtgaagccttcggagaccctgtccctcacctgcactgtctct, tggagctggatccggcagcccccagggaagggactgGAGTGGATTGGGTAT, tggagctggatccggcagcccccagggaagggactgGAGTGGATTGGGAAT, AACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCATTAGACACGTCCAAGAACCAGTTCTCCCTGAACCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTTT, AACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCCTTAGACACGTCCAAGAACCAGTTCTCCCTGACCCTGAGGTCTGTGACCGCTGCGGACACGGCCGTGTATTTT, GGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GGCCAGGGAACCCTGGTCACCGTCTCCTCAG, QVQLQESGPGLVKPSETLSLTCTVSGGSVSSGGYYWSWIRQPPGKGLEWIGYIYYSGSTNYNPSLKSRVSISLDTSKNQFSLNLSSVTAADTAVYF, QVQLQESGPGLVKPSETLSLTCTVSGGSVSSGGYYWSWIRQPPGKGLEWIGNIYYSGTTNYNPSLKSRVSISLDTSKNQFSLTLRSVTAADTAVYF, GQGTLV~SPSPQ, GQGTLV~SPSPQ
#> 3                           GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCACCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTGAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAAGCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGAATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCAGCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTAAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAACCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, 416, 792, 21, 777, IGHV5-51*01, IGHV5-51*01, IGHJ4*01, IGHJ4*01, ggatacagctttaccagctactgg, ggatacagctttaccagctactgg, ATCTATCCTGGTGACTCTGATACC, ATCTATCCTGGTGACTCTGATACC, TGTGTGAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAAGCTTTGACTACTGG, TGTGTAAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAACCTTTGACTACTGG, gaggtgcagctggtgcagtctggagcagaggtgaaaaagcccggggagtctctgaagatctcctgtaagggttct, gaggtgcagctggtgcagtctggagcagaggtgaaaaagcccggggagtctctgaagatctcctgtaagggttct, atcggctgggtgcgccagatgcccgggaaaggcctgGAGTGGATGGGGATC, atcggctgggtgcgccagatgcccgggaaaggcctgGAGTGGATGGGAATC, AGATACAGCCCGTCCTTCCAAGACCAGGTCACCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTAC, AGATACAGCCCGTCCTTCCAAGACCAGGTCAGCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTAC, GGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GGCCAGGGAACCCTGGTCACCGTCTCCTCAG, EVQLVQSGAEVKKPGESLKISCKGSGYSFTSYWIGWVRQMPGKGLEWMGIIYPGDSDTRYSPSFQDQVTMSVGKSIGTAYLQWSSLKASDTAMYY, EVQLVQSGAEVKKPGESLKISCKGSGYSFTSYWIGWVRQMPGKGLEWMGIIYPGDSDTRYSPSFQDQVSMSVGKSIGTAYLQWSSLKASDTAMYY, GQGTLV~SPSPQ, GQGTLV~SPSPQ
#> 4                                                                                       GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGAAAAGGCCTGGAGTGGGTGGGATCGATCTATCCTCGTGACTCTGATATCAGATACAATCCGTCCTTCCAAGGCCAAGTCACCATCTCAGCCGACAGGTCCACCACCACCGCCTACCTGCAGTGGAGCAGTCTGAAGGCCTCGGACACCGCCATATATTACTGTACGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAGCCCTGGTCACCGTCTCCTCAG, GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGTTGGGGAGCATCTACCCTCGTGACTCTGACACTAGATACAGTCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCCTCAGTACCGCCTACTTGCAGTGGAACAGCCTGCAGGCCTCGGACACCGCCATCTACTACTGTGCGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, 283, 987, 184, 57, IGHV5-51*01, IGHV5-51*01, IGHJ5*01, IGHJ5*01, ggatacagctttaccagctactgg, ggatacagctttaccagctactgg, ATCTATCCTCGTGACTCTGATATC, ATCTACCCTCGTGACTCTGACACT, TGTACGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGG, TGTGCGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGG, gaggtgcagctggtgcagtctggagcagaggtgaaaaagcccggggagtctctgaagatctcctgtaagggttct, gaggtgcagctggtgcagtctggagcagaggtgaaaaagcccggggagtctctgaagatctcctgtaagggttct, atcggctgggtgcgccagatgCCCGGAAAAGGCCTGGAGTGGGTGGGATCG, atcggctgggtgcgccagatgCCCGGGAAAGGCCTGGAGTGGTTGGGGAGC, AGATACAATCCGTCCTTCCAAGGCCAAGTCACCATCTCAGCCGACAGGTCCACCACCACCGCCTACCTGCAGTGGAGCAGTCTGAAGGCCTCGGACACCGCCATATATTAC, AGATACAGTCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCCTCAGTACCGCCTACTTGCAGTGGAACAGCCTGCAGGCCTCGGACACCGCCATCTACTAC, GGCCAGGGAGCCCTGGTCACCGTCTCCTCAG, GGCCAGGGAACCCTGGTCACCGTCTCCTCAG, EVQLVQSGAEVKKPGESLKISCKGSGYSFTSYWIGWVRQMPGKGLEWVGSIYPRDSDIRYNPSFQGQVTISADRSTTTAYLQWSSLKASDTAIYY, EVQLVQSGAEVKKPGESLKISCKGSGYSFTSYWIGWVRQMPGKGLEWLGSIYPRDSDTRYSPSFQGQVTISADKSLSTAYLQWNSLQASDTAIYY, GQGALV~SPSPQ, GQGTLV~SPSPQ
#>