This function creates germlines for clonal lineages. B cell clonal lineage represents a set of B cells that presumably have a common origin (arising from the same VDJ rearrangement event) and a common ancestor. Each clonal lineage has its own germline sequence that represents the ancestral sequence for each BCR in clonal lineage. In other words, germline sequence is a sequence of B-cells immediately after VDJ recombination, before B-cell maturation and hypermutation process. Germline sequence is useful for assessing the degree of mutation and maturity of the repertoire.

Creates germlines for clonal lineages

repGermline(.data, species, min_nuc_outside_cdr3, ref_only_first)

Arguments

.data
The data to be processed. Can be data.frame, data.table or a list of these objects.

It must have columns in the immunarch compatible format immunarch_data_format.

species
Species from which the data was acquired. Available options: "HomoSapiens" (default), "MusMusculus", "BosTaurus", "CamelusDromedarius", "CanisLupusFamiliaris", "DanioRerio", "MacacaMulatta", "MusMusculusDomesticus", "MusMusculusCastaneus", "MusMusculusMolossinus", "MusMusculusMusculus", "MusSpretus", "OncorhynchusMykiss", "OrnithorhynchusAnatinus", "OryctolagusCuniculus", "RattusNorvegicus", "SusScrofa".

min_nuc_outside_cdr3
This parameter sets how many nucleotides should have V or J chain outside of CDR3 to be considered good for further alignment.

ref_only_first
This parameter, if TRUE, means to take only first sequence from reference for each allele name; if FALSE, all sequences will be taken, and output table will increase in size as a result.

Value

Data with added columns V.first.allele, J.first.allele (with first alleles of V and J genes), V.sequence, J.sequence (with V and J reference sequences), Germline.sequence (with combined germline sequence)

Examples

data(bcrdata)

bcrdata$data %>%
  repGermline()
#> Warning: Alleles IGHV4-55*01 from sample full_clones not found in the reference and will be dropped!
#> Probably, species argument is wrong (current value: HomoSapiens) or the data contains non-BCR genes.
#> $full_clones
#> # A tibble: 999 × 39
#> # Rowwise: 
#>    J.first.allele V.first.allele Clones Proportion CDR3.nt CDR3.aa V.name D.name
#>    <chr>          <chr>           <dbl>      <dbl> <chr>   <chr>   <chr>  <chr> 
#>  1 IGHJ1*01       IGHV2-5*01         24    2.21e-6 TGTGTC… CVEGQP… IGHV2… IGHD6…
#>  2 IGHJ1*01       IGHV4-30-2*01      59    5.42e-6 TGTGTG… CVRDQR… IGHV4… IGHD2…
#>  3 IGHJ1*01       IGHV1-3*01          8    7.35e-7 TGTGCG… CARGIT… IGHV1… IGHD3…
#>  4 IGHJ1*01       IGHV1-69*01       824    7.57e-5 TGTGCG… CATPAE… IGHV1… IGHD5…
#>  5 IGHJ1*01       IGHV1-69*01        42    3.86e-6 TGTGTG… CVRDSA… IGHV1… IGHD6…
#>  6 IGHJ2*01       IGHV1-46*01         7    6.43e-7 TGTGCG… CARASR… IGHV1… IGHD5…
#>  7 IGHJ2*01       IGHV1-18*01        21    1.93e-6 TGTGCG… CARSYY… IGHV1… IGHD3…
#>  8 IGHJ2*01       IGHV3-21*01         3    2.76e-7 TGTGCA… CARVPS… IGHV3… IGHD3…
#>  9 IGHJ3*01       IGHV1-18*01        26    2.39e-6 TGTGCG… CARVLT… IGHV1… IGHD3…
#> 10 IGHJ3*01       IGHV1-2*01         10    9.19e-7 TGTGCG… CARGGG… IGHV1… IGHD3…
#> # … with 989 more rows, and 31 more variables: J.name <chr>, V.end <int>,
#> #   D.start <int>, D.end <int>, J.start <int>, VJ.ins <int>, VD.ins <int>,
#> #   DJ.ins <int>, Sequence <chr>, CDR3.start <chr>, CDR3.end <chr>,
#> #   C.name <chr>, C.start <list>, C.end <list>, CDR1.nt <chr>, CDR1.aa <chr>,
#> #   CDR2.nt <chr>, CDR2.aa <chr>, FR1.nt <chr>, FR1.aa <chr>, FR2.nt <chr>,
#> #   FR2.aa <chr>, FR3.nt <chr>, FR3.aa <chr>, FR4.nt <chr>, FR4.aa <chr>,
#> #   V3.Deletions <chr>, J3.Deletions <chr>, V.sequence <chr>, …
#>