Save immune repertoires to the disk

The repSave function is deigned to save your data to the disk in desirable format. Currently supports "immunarch" and "vdjtools" file formats.

repSave(.data, .path, .format = c("immunarch", "vdjtools"), .compress = TRUE)

Arguments

.data

An R dataframe, a list of R dataframes or a list with data and meta where first element is a list of dataframes and the latter is a dataframe with metadata.

.path

A string with the path to the output directory. It should include file name if a single dataframe is provided to .data argument.

.format

A string with desirable format specification. Current options are "immunarch" and "vdjtools".

.compress

A boolean value. Defines whether the output will be compressed or not.

Value

No return value.

Details

It is not necessary to create directories beforehand. If the provided directory does not exist it will be created automatically.

Examples

data(immdata)
# Reduce data to save time on examples
immdata$data <- purrr::map(immdata$data, ~ .x %>% head(10))
dirpath <- tempdir()
# Save the list of repertoires
repSave(immdata, dirpath)
#> Warning: The `path` argument of `write_tsv()` is deprecated as of readr 1.4.0.
#> Please use the `file` argument instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
#> Warning: The `path` argument of `write_lines()` is deprecated as of readr 1.4.0.
#> Please use the `file` argument instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
# Load it and check if it is the same
new_immdata <- repLoad(dirpath)
#> 
#> == Step 1/3: loading repertoire files... ==
#> Processing "/tmp/Rtmp14wr6E" ...
#>   -- [1/29] Parsing "/tmp/Rtmp14wr6E/A2-i129.tsv.gz" -- 
#> immunarch
#>   -- [2/29] Parsing "/tmp/Rtmp14wr6E/A2-i131.tsv.gz" -- 
#> immunarch
#>   -- [3/29] Parsing "/tmp/Rtmp14wr6E/A2-i132.tsv.gz" -- 
#> immunarch
#>   -- [4/29] Parsing "/tmp/Rtmp14wr6E/A2-i133.tsv.gz" -- 
#> immunarch
#>   -- [5/29] Parsing "/tmp/Rtmp14wr6E/A4-i191.tsv.gz" -- 
#> immunarch
#>   -- [6/29] Parsing "/tmp/Rtmp14wr6E/A4-i192.tsv.gz" -- 
#> immunarch
#>   -- [7/29] Parsing "/tmp/Rtmp14wr6E/MS1.tsv.gz" -- 
#> immunarch
#>   -- [8/29] Parsing "/tmp/Rtmp14wr6E/MS2.tsv.gz" -- 
#> immunarch
#>   -- [9/29] Parsing "/tmp/Rtmp14wr6E/MS3.tsv.gz" -- 
#> immunarch
#>   -- [10/29] Parsing "/tmp/Rtmp14wr6E/MS4.tsv.gz" -- 
#> immunarch
#>   -- [11/29] Parsing "/tmp/Rtmp14wr6E/MS5.tsv.gz" -- 
#> immunarch
#>   -- [12/29] Parsing "/tmp/Rtmp14wr6E/MS6.tsv.gz" -- 
#> immunarch
#>   -- [13/29] Parsing "/tmp/Rtmp14wr6E/clustal21d5e258dd6e9.dnd" -- 
#> unsupported format, skipping
#>   -- [14/29] Parsing "/tmp/Rtmp14wr6E/clustal21d5e7840a494.dnd" -- 
#> unsupported format, skipping
#>   -- [15/29] Parsing "/tmp/Rtmp14wr6E/clustal21d5f258dd6e9.dnd" -- 
#> unsupported format, skipping
#>   -- [16/29] Parsing "/tmp/Rtmp14wr6E/clustal21d5f7840a494.dnd" -- 
#> unsupported format, skipping
#>   -- [17/29] Parsing "/tmp/Rtmp14wr6E/clustal21db0258dd6e9.dnd" -- 
#> unsupported format, skipping
#>   -- [18/29] Parsing "/tmp/Rtmp14wr6E/clustal21db048cf8583.dnd" -- 
#> unsupported format, skipping
#>   -- [19/29] Parsing "/tmp/Rtmp14wr6E/clustal21db1258dd6e9.dnd" -- 
#> unsupported format, skipping
#>   -- [20/29] Parsing "/tmp/Rtmp14wr6E/clustal21db148cf8583.dnd" -- 
#> unsupported format, skipping
#>   -- [21/29] Parsing "/tmp/Rtmp14wr6E/file2191313b50ca6" -- 
#> unsupported format, skipping
#>   -- [22/29] Parsing "/tmp/Rtmp14wr6E/file2191326c99ce0" -- 
#> unsupported format, skipping
#>   -- [23/29] Parsing "/tmp/Rtmp14wr6E/file2191344ea1cda" -- 
#> unsupported format, skipping
#>   -- [24/29] Parsing "/tmp/Rtmp14wr6E/file219134c4de9bd" -- 
#> unsupported format, skipping
#>   -- [25/29] Parsing "/tmp/Rtmp14wr6E/file21913568b5c96" -- 
#> unsupported format, skipping
#>   -- [26/29] Parsing "/tmp/Rtmp14wr6E/file2191357a6872f" -- 
#> unsupported format, skipping
#>   -- [27/29] Parsing "/tmp/Rtmp14wr6E/file2191367c572ae" -- 
#> unsupported format, skipping
#>   -- [28/29] Parsing "/tmp/Rtmp14wr6E/file219137517182d" -- 
#> unsupported format, skipping
#>   -- [29/29] Parsing "/tmp/Rtmp14wr6E/metadata.txt" -- 
#> metadata
#> Processing "/tmp/Rtmp14wr6E/downlit" ...
#>   -- [1/13] Parsing "/tmp/Rtmp14wr6E/downlit/MASS" -- 
#> unsupported format, skipping
#>   -- [2/13] Parsing "/tmp/Rtmp14wr6E/downlit/Rtsne" -- 
#> unsupported format, skipping
#>   -- [3/13] Parsing "/tmp/Rtmp14wr6E/downlit/base" -- 
#> unsupported format, skipping
#>   -- [4/13] Parsing "/tmp/Rtmp14wr6E/downlit/data.table" -- 
#> unsupported format, skipping
#>   -- [5/13] Parsing "/tmp/Rtmp14wr6E/downlit/devtools" -- 
#> unsupported format, skipping
#>   -- [6/13] Parsing "/tmp/Rtmp14wr6E/downlit/dplyr" -- 
#> unsupported format, skipping
#>   -- [7/13] Parsing "/tmp/Rtmp14wr6E/downlit/factoextra" -- 
#> unsupported format, skipping
#>   -- [8/13] Parsing "/tmp/Rtmp14wr6E/downlit/fpc" -- 
#> unsupported format, skipping
#>   -- [9/13] Parsing "/tmp/Rtmp14wr6E/downlit/magrittr" -- 
#> unsupported format, skipping
#>   -- [10/13] Parsing "/tmp/Rtmp14wr6E/downlit/pkgload" -- 
#> unsupported format, skipping
#>   -- [11/13] Parsing "/tmp/Rtmp14wr6E/downlit/remotes" -- 
#> unsupported format, skipping
#>   -- [12/13] Parsing "/tmp/Rtmp14wr6E/downlit/stats" -- 
#> unsupported format, skipping
#>   -- [13/13] Parsing "/tmp/Rtmp14wr6E/downlit/utils" -- 
#> unsupported format, skipping
#> 
#> == Step 2/3: checking metadata files and merging files... ==
#> Processing "/tmp/Rtmp14wr6E" ...
#>   -- Everything is OK!
#> Processing "/tmp/Rtmp14wr6E/downlit" ...
#>   -- Metadata file not found; creating a dummy metadata...
#> Warning: Unknown or uninitialised column: `Sample`.
#> Warning: Unknown or uninitialised column: `Sample`.
#> 
#> == Step 3/3: processing paired chain data... ==
#> Done!
# sum(immdata$data[[1]] != new_immdata$data[[1]], na.rm = TRUE)
# sum(immdata$data[[2]] != new_immdata$data[[2]], na.rm = TRUE)
# sum(immdata$meta != new_immdata$meta, na.rm = TRUE)