Save immune repertoires to the disk

The repSave function is deigned to save your data to the disk in desirable format. Currently supports "immunarch" and "vdjtools" file formats.

repSave(.data, .path, .format = c("immunarch", "vdjtools"), .compress = TRUE)

Arguments

.data
An R dataframe, a list of R dataframes or a list with data and meta where first element is a list of dataframes and the latter is a dataframe with metadata.

.path
A string with the path to the output directory. It should include file name if a single dataframe is provided to .data argument.

.format
A string with desirable format specification. Current options are "immunarch" and "vdjtools".

.compress
A boolean value. Defines whether the output will be compressed or not.

Value

No return value.

Details

It is not necessary to create directories beforehand. If the provided directory does not exist it will be created automatically.

Examples

data(immdata)
dirpath <- tempdir()
# Save the list of repertoires
repSave(immdata, dirpath)

#> Warning: The `path` argument of `write_tsv()` is deprecated as of readr 1.4.0.
#> Please use the `file` argument instead.

#> Warning: The `path` argument of `write_lines()` is deprecated as of readr 1.4.0.
#> Please use the `file` argument instead.

# Load it and check if it is the same
new_immdata <- repLoad(dirpath)

#> 
#> == Step 1/3: loading repertoire files... ==

#> Processing "/tmp/RtmpqsPKoZ" ...

#>   -- [1/15] Parsing "/tmp/RtmpqsPKoZ/A2-i129.tsv.gz" -- 

#> immunarch

#>   -- [2/15] Parsing "/tmp/RtmpqsPKoZ/A2-i131.tsv.gz" -- 

#> immunarch

#>   -- [3/15] Parsing "/tmp/RtmpqsPKoZ/A2-i132.tsv.gz" -- 

#> immunarch

#>   -- [4/15] Parsing "/tmp/RtmpqsPKoZ/A2-i133.tsv.gz" -- 

#> immunarch

#>   -- [5/15] Parsing "/tmp/RtmpqsPKoZ/A4-i191.tsv.gz" -- 

#> immunarch

#>   -- [6/15] Parsing "/tmp/RtmpqsPKoZ/A4-i192.tsv.gz" -- 

#> immunarch

#>   -- [7/15] Parsing "/tmp/RtmpqsPKoZ/file18dfda2185cb8c" -- 

#> unsupported format, skipping

#>   -- [8/15] Parsing "/tmp/RtmpqsPKoZ/file18dfda5a3170fd" -- 

#> unsupported format, skipping

#>   -- [9/15] Parsing "/tmp/RtmpqsPKoZ/metadata.txt" -- 

#> metadata

#>   -- [10/15] Parsing "/tmp/RtmpqsPKoZ/MS1.tsv.gz" -- 

#> immunarch

#>   -- [11/15] Parsing "/tmp/RtmpqsPKoZ/MS2.tsv.gz" -- 

#> immunarch

#>   -- [12/15] Parsing "/tmp/RtmpqsPKoZ/MS3.tsv.gz" -- 

#> immunarch

#>   -- [13/15] Parsing "/tmp/RtmpqsPKoZ/MS4.tsv.gz" -- 

#> immunarch

#>   -- [14/15] Parsing "/tmp/RtmpqsPKoZ/MS5.tsv.gz" -- 

#> immunarch

#>   -- [15/15] Parsing "/tmp/RtmpqsPKoZ/MS6.tsv.gz" -- 

#> immunarch

#> Processing "/tmp/RtmpqsPKoZ/downlit" ...

#>   -- [1/8] Parsing "/tmp/RtmpqsPKoZ/downlit/base" -- 

#> unsupported format, skipping

#>   -- [2/8] Parsing "/tmp/RtmpqsPKoZ/downlit/devtools" -- 

#> unsupported format, skipping

#>   -- [3/8] Parsing "/tmp/RtmpqsPKoZ/downlit/factoextra" -- 

#> unsupported format, skipping

#>   -- [4/8] Parsing "/tmp/RtmpqsPKoZ/downlit/fpc" -- 

#> unsupported format, skipping

#>   -- [5/8] Parsing "/tmp/RtmpqsPKoZ/downlit/MASS" -- 

#> unsupported format, skipping

#>   -- [6/8] Parsing "/tmp/RtmpqsPKoZ/downlit/Rtsne" -- 

#> unsupported format, skipping

#>   -- [7/8] Parsing "/tmp/RtmpqsPKoZ/downlit/stats" -- 

#> unsupported format, skipping

#>   -- [8/8] Parsing "/tmp/RtmpqsPKoZ/downlit/utils" -- 

#> unsupported format, skipping

#> 
#> == Step 2/3: checking metadata files and merging files... ==

#> Processing "/tmp/RtmpqsPKoZ" ...

#>   -- Everything is OK!

#> Processing "/tmp/RtmpqsPKoZ/downlit" ...

#>   -- Metadata file not found; creating a dummy metadata...

#> Warning: Unknown or uninitialised column: `Sample`.

#> Warning: Unknown or uninitialised column: `Sample`.

#> 
#> == Step 3/3: processing paired chain data... ==

#> Done!

# sum(immdata$data[[1]] != new_immdata$data[[1]], na.rm = TRUE)
# sum(immdata$data[[2]] != new_immdata$data[[2]], na.rm = TRUE)
# sum(immdata$meta != new_immdata$meta, na.rm = TRUE)