Calculates number of mutations against the germline for each clonotype

This function aligns V and J genes from the germline in each cluster with corresponding genes in each clonotype, saves the alignments for purpose of visualization, and calculates number of mutations for each clonotype.

repSomaticHypermutation(.data, .threads, .nofail)

Arguments

.data

The data to be processed: an output of repClonalFamily(); variants with one sample and list of samples are both supported.

.threads

Number of threads to use.

.nofail

Will return NA instead of stopping if Clustal W is not installed. Used to avoid raising errors in examples on computers where Clustal W is not installed.

Value

Dataframe or list of dataframes (if input is a list with multiple samples). The dataframe has all the columns from repClonalFamily() output dataframe, with Sequence column unnested: the resulting dataframe has one line per clonotype. Clone.ID column contains original IDs for clonotypes, and can be used as dataframe key. New columns are added: * Germline.Alignment.V: contains V gene alignment of current clonotype with the germline * Germline.Alignment.J: contains J gene alignment of current clonotype with the germline * Substitutions: contains number of substitutions in the alignment (summary for V and J) * Insertions: contains number of insertions in the clonotype relative to germline (summary for V and J) * Deletions: contains number of deletions in the clonotype relative to germline (summary for V and J) * Mutations: contains total number of mutations in the alignment (summary for V and J)

Examples

data(bcrdata)
bcr_data <- bcrdata$data

bcr_data %>%
  seqCluster(seqDist(bcr_data), .fixed_threshold = 3) %>%
  repGermline(.threads = 1) %>%
  repAlignLineage(.min_lineage_sequences = 2, .align_threads = 2, .nofail = TRUE) %>%
  repClonalFamily(.threads = 1, .nofail = TRUE) %>%
  repSomaticHypermutation(.threads = 1, .nofail = TRUE)
#> Warning: Alleles IGHV4-55*01 from sample full_clones not found in the reference and will be dropped!
#> Probably, species argument is wrong (current value: HomoSapiens) or the data contains non-BCR genes.
#> $full_clones
#>                              Cluster
#> 1 IGHV4-61/IGHJ4_length_45_cluster_1
#> 2 IGHV4-61/IGHJ4_length_45_cluster_1
#> 3 IGHV4-61/IGHJ4_length_63_cluster_1
#> 4 IGHV4-61/IGHJ4_length_63_cluster_1
#> 5 IGHV5-51/IGHJ4_length_60_cluster_1
#> 6 IGHV5-51/IGHJ4_length_60_cluster_1
#> 7 IGHV5-51/IGHJ5_length_45_cluster_1
#> 8 IGHV5-51/IGHJ5_length_45_cluster_1
#>                                                                                                                                                                                                                                                                                                                                                                                   Germline.Input
#> 1                   CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 2                   CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 3 CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 4 CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 5       GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 6       GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 7                      GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 8                      GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#>                                                                                                                                                                                                                                                                                      V.germline.nt
#> 1 CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTAC
#> 2 CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTAC
#> 3 CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTAC
#> 4 CAGGTGCAGCTGCAGGAGTCGGGCCCAGGACTGGTGAAGCCTTCGGAGACCCTGTCCCTCACCTGCACTGTCTCTGGTGGCTCCGTCAGCAGTGGTAGTTACTACTGGAGCTGGATCCGGCAGCCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTAC
#> 5    GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTAC
#> 6    GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTAC
#> 7    GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTAC
#> 8    GAGGTGCAGCTGGTGCAGTCTGGAGCAGAGGTGAAAAAGCCCGGGGAGTCTCTGAAGATCTCCTGTAAGGGTTCTGGATACAGCTTTACCAGCTACTGGATCGGCTGGGTGCGCCAGATGCCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTAC
#>                     J.germline.nt CDR3.germline.length V.length J.length
#> 1 GGCCAAGGAACCCTGGTCACCGTCTCCTCAG                   45      165       31
#> 2 GGCCAAGGAACCCTGGTCACCGTCTCCTCAG                   45      165       31
#> 3 GGCCAAGGAACCCTGGTCACCGTCTCCTCAG                   63      147       31
#> 4 GGCCAAGGAACCCTGGTCACCGTCTCCTCAG                   63      147       31
#> 5 GGCCAAGGAACCCTGGTCACCGTCTCCTCAG                   60      150       31
#> 6 GGCCAAGGAACCCTGGTCACCGTCTCCTCAG                   60      150       31
#> 7 GGCCAAGGAACCCTGGTCACCGTCTCCTCAG                   45      165       31
#> 8 GGCCAAGGAACCCTGGTCACCGTCTCCTCAG                   45      165       31
#>                                                                                                                                                                                                                                     Germline.Output
#> 1 CCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 2 CCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 3 GAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 4 GAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 5 GAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 6 GAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 7 CCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 8 CCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#>                                                                                                                                                                                                                                     Common.Ancestor
#> 1 CCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCCATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATGTCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGTTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACRSCACTCCCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#> 2 CCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCCATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATGTCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGTTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACRSCACTCCCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#> 3 GAGTGGATTGGGTATATCTATTACAGTGGGAGTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCATTAGACACGTCCAAGAACCAGTTCTCCCTGAACCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACWAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#> 4 GAGTGGATTGGGTATATCTATTACAGTGGGAGTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCATTAGACACGTCCAAGAACCAGTTCTCCCTGAACCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACWAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#> 5 GAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCACCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTRAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAASCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#> 6 GAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCACCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTRAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAASCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#> 7 CCCGGGAAAGGCCTGGAGTGGDTGGGGABCATCTATCCTCGTGACTCTGATACCAGATACAGTCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATVTATTACTGTRCGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#> 8 CCCGGGAAAGGCCTGGAGTGGDTGGGGABCATCTATCCTCGTGACTCTGATACCAGATACAGTCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATVTATTACTGTRCGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#>   Trunk.Length
#> 1            6
#> 2            6
#> 3            7
#> 4            7
#> 5            6
#> 6            6
#> 7            6
#> 8            6
#>                                                                       Tree
#> 1   4, 4, 4, 1, 2, 3, 0.04149, 0.04979, 0.0249, 1, ID_643, ID_42, Germline
#> 2   4, 4, 4, 1, 2, 3, 0.04149, 0.04979, 0.0249, 1, ID_643, ID_42, Germline
#> 3 4, 4, 4, 1, 2, 3, 0.02282, 0.00207, 0.02905, 1, ID_123, ID_119, Germline
#> 4 4, 4, 4, 1, 2, 3, 0.02282, 0.00207, 0.02905, 1, ID_123, ID_119, Germline
#> 5  4, 4, 4, 1, 2, 3, 0.01245, 0.00415, 0.0249, 1, ID_792, ID_416, Germline
#> 6  4, 4, 4, 1, 2, 3, 0.01245, 0.00415, 0.0249, 1, ID_792, ID_416, Germline
#> 7 4, 4, 4, 1, 2, 3, 0.04772, 0.06017, 0.02075, 1, ID_987, ID_283, Germline
#> 8 4, 4, 4, 1, 2, 3, 0.04772, 0.06017, 0.02075, 1, ID_987, ID_283, Germline
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            TreeStats
#> 1  CommonAncestor_1, ID_643, ID_42, Germline, CommonAncestor, Clonotype, Clonotype, Germline, 1, 301, 26, 1, Germline, CommonAncestor_1, CommonAncestor_1, NA, 6, 15, 17, 0, 4, 9, 12, 0, CCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCCATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATGTCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGTTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACRSCACTCCCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, CCCCCAGGGAAGGGACTGGAGTGGATTGGATATATTCATTACAGTGGGAACTCCTTCTACAACCCCTCCCTCAAAAGTCGAGTCACCATGTCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAGGTTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACGCCACTCCCTTTGACTACTGGGGCCCGGGAACCCTGGTCACCGTCTCCTCAG, CCCCCAGGGAAGGGACTGGAGTGGATTGGGTATGTCCATCACAGTGGGAGCACCAATTACAATCCCTCCCTCAAGCGTCGAGTCACCATGTCAGTAGACACGGCCAGGAAGCAGTTGTCCCTGAAATTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACAGCACTCCCTTTGACTACTGGGGCCAGGGAGCCCTGGTCACCGTCTCCTCAG, CCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 2  CommonAncestor_1, ID_643, ID_42, Germline, CommonAncestor, Clonotype, Clonotype, Germline, 1, 301, 26, 1, Germline, CommonAncestor_1, CommonAncestor_1, NA, 6, 15, 17, 0, 4, 9, 12, 0, CCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCCATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATGTCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGTTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACRSCACTCCCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, CCCCCAGGGAAGGGACTGGAGTGGATTGGATATATTCATTACAGTGGGAACTCCTTCTACAACCCCTCCCTCAAAAGTCGAGTCACCATGTCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAGGTTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACGCCACTCCCTTTGACTACTGGGGCCCGGGAACCCTGGTCACCGTCTCCTCAG, CCCCCAGGGAAGGGACTGGAGTGGATTGGGTATGTCCATCACAGTGGGAGCACCAATTACAATCCCTCCCTCAAGCGTCGAGTCACCATGTCAGTAGACACGGCCAGGAAGCAGTTGTCCCTGAAATTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACAGCACTCCCTTTGACTACTGGGGCCAGGGAGCCCTGGTCACCGTCTCCTCAG, CCCCCAGGGAAGGGACTGGAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 3   CommonAncestor_1, ID_123, ID_119, Germline, CommonAncestor, Clonotype, Clonotype, Germline, 1, 53, 290, 1, Germline, CommonAncestor_1, CommonAncestor_1, NA, 7, 12, 7, 0, 4, 7, 4, 0, GAGTGGATTGGGTATATCTATTACAGTGGGAGTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCATTAGACACGTCCAAGAACCAGTTCTCCCTGAACCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACWAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATTGGGAATATCTATTACAGTGGGACTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCCTTAGACACGTCCAAGAACCAGTTCTCCCTGACCCTGAGGTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACAAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATTGGGTATATCTATTACAGTGGGAGTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCATTAGACACGTCCAAGAACCAGTTCTCCCTGAACCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACTAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 4   CommonAncestor_1, ID_123, ID_119, Germline, CommonAncestor, Clonotype, Clonotype, Germline, 1, 53, 290, 1, Germline, CommonAncestor_1, CommonAncestor_1, NA, 7, 12, 7, 0, 4, 7, 4, 0, GAGTGGATTGGGTATATCTATTACAGTGGGAGTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCATTAGACACGTCCAAGAACCAGTTCTCCCTGAACCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACWAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATTGGGAATATCTATTACAGTGGGACTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCCTTAGACACGTCCAAGAACCAGTTCTCCCTGACCCTGAGGTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACAAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATTGGGTATATCTATTACAGTGGGAGTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCATTAGACACGTCCAAGAACCAGTTCTCCCTGAACCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACTAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATTGGGTATATCTATTACAGTGGGAGCACCAACTACAACCCCTCCCTCAAGAGTCGAGTCACCATATCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAAGCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 5    CommonAncestor_1, ID_792, ID_416, Germline, CommonAncestor, Clonotype, Clonotype, Germline, 1, 777, 21, 1, Germline, CommonAncestor_1, CommonAncestor_1, NA, 6, 8, 6, 0, 5, 6, 5, 0, GAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCACCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTRAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAASCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATGGGAATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCAGCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTAAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAACCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCACCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTGAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAAGCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 6    CommonAncestor_1, ID_792, ID_416, Germline, CommonAncestor, Clonotype, Clonotype, Germline, 1, 777, 21, 1, Germline, CommonAncestor_1, CommonAncestor_1, NA, 6, 8, 6, 0, 5, 6, 5, 0, GAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCACCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTRAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAASCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATGGGAATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCAGCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTAAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAACCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCACCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTGAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAAGCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, GAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 7 CommonAncestor_1, ID_987, ID_283, Germline, CommonAncestor, Clonotype, Clonotype, Germline, 1, 57, 184, 1, Germline, CommonAncestor_1, CommonAncestor_1, NA, 6, 15, 18, 0, 4, 7, 10, 0, CCCGGGAAAGGCCTGGAGTGGDTGGGGABCATCTATCCTCGTGACTCTGATACCAGATACAGTCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATVTATTACTGTRCGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, CCCGGGAAAGGCCTGGAGTGGTTGGGGAGCATCTACCCTCGTGACTCTGACACTAGATACAGTCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCCTCAGTACCGCCTACTTGCAGTGGAACAGCCTGCAGGCCTCGGACACCGCCATCTACTACTGTGCGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, CCCGGAAAAGGCCTGGAGTGGGTGGGATCGATCTATCCTCGTGACTCTGATATCAGATACAATCCGTCCTTCCAAGGCCAAGTCACCATCTCAGCCGACAGGTCCACCACCACCGCCTACCTGCAGTGGAGCAGTCTGAAGGCCTCGGACACCGCCATATATTACTGTACGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAGCCCTGGTCACCGTCTCCTCAG, CCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#> 8 CommonAncestor_1, ID_987, ID_283, Germline, CommonAncestor, Clonotype, Clonotype, Germline, 1, 57, 184, 1, Germline, CommonAncestor_1, CommonAncestor_1, NA, 6, 15, 18, 0, 4, 7, 10, 0, CCCGGGAAAGGCCTGGAGTGGDTGGGGABCATCTATCCTCGTGACTCTGATACCAGATACAGTCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATVTATTACTGTRCGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, CCCGGGAAAGGCCTGGAGTGGTTGGGGAGCATCTACCCTCGTGACTCTGACACTAGATACAGTCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCCTCAGTACCGCCTACTTGCAGTGGAACAGCCTGCAGGCCTCGGACACCGCCATCTACTACTGTGCGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAG, CCCGGAAAAGGCCTGGAGTGGGTGGGATCGATCTATCCTCGTGACTCTGATATCAGATACAATCCGTCCTTCCAAGGCCAAGTCACCATCTCAGCCGACAGGTCCACCACCACCGCCTACCTGCAGTGGAGCAGTCTGAAGGCCTCGGACACCGCCATATATTACTGTACGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAGCCCTGGTCACCGTCTCCTCAG, CCCGGGAAAGGCCTGGAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCATCAGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGGCCAAGGAACCCTGGTCACCGTCTCCTCAG
#>                                                                                                                                                                                                                                                 Sequence
#> 1  CCCCCAGGGAAGGGACTGGAGTGGATTGGGTATGTCCATCACAGTGGGAGCACCAATTACAATCCCTCCCTCAAGCGTCGAGTCACCATGTCAGTAGACACGGCCAGGAAGCAGTTGTCCCTGAAATTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACAGCACTCCCTTTGACTACTGGGGCCAGGGAGCCCTGGTCACCGTCTCCTCAGGTAA
#> 2 CCCCCAGGGAAGGGACTGGAGTGGATTGGATATATTCATTACAGTGGGAACTCCTTCTACAACCCCTCCCTCAAAAGTCGAGTCACCATGTCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAGGTTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTCTGTGCGAGAGGCGCCGGGGACTACGCCACTCCCTTTGACTACTGGGGCCCGGGAACCCTGGTCACCGTCTCCTCAGGTAAG
#> 3 GAGTGGATTGGGTATATCTATTACAGTGGGAGTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCATTAGACACGTCCAAGAACCAGTTCTCCCTGAACCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACTAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAGGTAAG
#> 4 GAGTGGATTGGGAATATCTATTACAGTGGGACTACCAACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCCTTAGACACGTCCAAGAACCAGTTCTCCCTGACCCTGAGGTCTGTGACCGCTGCGGACACGGCCGTGTATTTTTGTGCGAGAGATCATCACAGTTTCTATGATACAAGTGGTTATTATCCCTACTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAGGTAAG
#> 5   GAGTGGATGGGGATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCACCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTGAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAAGCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAGGTA
#> 6 GAGTGGATGGGAATCATCTATCCTGGTGACTCTGATACCAGATACAGCCCGTCCTTCCAAGACCAGGTCAGCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTACTGTGTAAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAACCTTTGACTACTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAGGTAAG
#> 7 CCCGGAAAAGGCCTGGAGTGGGTGGGATCGATCTATCCTCGTGACTCTGATATCAGATACAATCCGTCCTTCCAAGGCCAAGTCACCATCTCAGCCGACAGGTCCACCACCACCGCCTACCTGCAGTGGAGCAGTCTGAAGGCCTCGGACACCGCCATATATTACTGTACGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAGCCCTGGTCACCGTCTCCTCAGGTAAG
#> 8 CCCGGGAAAGGCCTGGAGTGGTTGGGGAGCATCTACCCTCGTGACTCTGACACTAGATACAGTCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCCTCAGTACCGCCTACTTGCAGTGGAACAGCCTGCAGGCCTCGGACACCGCCATCTACTACTGTGCGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGGGGCCAGGGAACCCTGGTCACCGTCTCCTCAGGTAAG
#>   Clone.ID Clones                        CDR1.nt                  CDR2.nt
#> 1       42     26 ggtggctccgtcagcagtggtggttactac    GTCCATCACAGTGGGAGCACC
#> 2      643    301 ggtggctccgtcagcagtggtggttactac    ATTCATTACAGTGGGAACTCC
#> 3      119    290 ggtggctccgtcagcagtggtggttactac    ATCTATTACAGTGGGAGTACC
#> 4      123     53 ggtggctccgtcagcagtggtggttactac    ATCTATTACAGTGGGACTACC
#> 5      416     21       ggatacagctttaccagctactgg ATCTATCCTGGTGACTCTGATACC
#> 6      792    777       ggatacagctttaccagctactgg ATCTATCCTGGTGACTCTGATACC
#> 7      283    184       ggatacagctttaccagctactgg ATCTATCCTCGTGACTCTGATATC
#> 8      987     57       ggatacagctttaccagctactgg ATCTACCCTCGTGACTCTGACACT
#>                                                           CDR3.nt
#> 1                   TGTGCGAGAGGCGCCGGGGACTACAGCACTCCCTTTGACTACTGG
#> 2                   TGTGCGAGAGGCGCCGGGGACTACGCCACTCCCTTTGACTACTGG
#> 3 TGTGCGAGAGATCATCACAGTTTCTATGATACTAGTGGTTATTATCCCTACTTTGACTACTGG
#> 4 TGTGCGAGAGATCATCACAGTTTCTATGATACAAGTGGTTATTATCCCTACTTTGACTACTGG
#> 5    TGTGTGAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAAGCTTTGACTACTGG
#> 6    TGTGTAAGACACTCTGGTAAGATAGAATCATCTGTGGGCCCCAGAACCTTTGACTACTGG
#> 7                   TGTACGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGG
#> 8                   TGTGCGAGACAACACCCTTTAGATAGTCGTGTCTGGTACACTTGG
#>                                                                        FR1.nt
#> 1 caggtgcagctgcaggagtcgggcccaggactggtgaagccttcggagaccctgtccctcacctgcactgtctct
#> 2 caggtgcagctgcaggagtcgggcccaggactggtgaagccttcggagaccctgtccctcacctgcactgtctct
#> 3 caggtgcagctgcaggagtcgggcccaggactggtgaagccttcggagaccctgtccctcacctgcactgtctct
#> 4 caggtgcagctgcaggagtcgggcccaggactggtgaagccttcggagaccctgtccctcacctgcactgtctct
#> 5 gaggtgcagctggtgcagtctggagcagaggtgaaaaagcccggggagtctctgaagatctcctgtaagggttct
#> 6 gaggtgcagctggtgcagtctggagcagaggtgaaaaagcccggggagtctctgaagatctcctgtaagggttct
#> 7 gaggtgcagctggtgcagtctggagcagaggtgaaaaagcccggggagtctctgaagatctcctgtaagggttct
#> 8 gaggtgcagctggtgcagtctggagcagaggtgaaaaagcccggggagtctctgaagatctcctgtaagggttct
#>                                                FR2.nt
#> 1 tggagctggatccggcagCCCCCAGGGAAGGGACTGGAGTGGATTGGGTAT
#> 2 tggagctggatccggcagCCCCCAGGGAAGGGACTGGAGTGGATTGGATAT
#> 3 tggagctggatccggcagcccccagggaagggactgGAGTGGATTGGGTAT
#> 4 tggagctggatccggcagcccccagggaagggactgGAGTGGATTGGGAAT
#> 5 atcggctgggtgcgccagatgcccgggaaaggcctgGAGTGGATGGGGATC
#> 6 atcggctgggtgcgccagatgcccgggaaaggcctgGAGTGGATGGGAATC
#> 7 atcggctgggtgcgccagatgCCCGGAAAAGGCCTGGAGTGGGTGGGATCG
#> 8 atcggctgggtgcgccagatgCCCGGGAAAGGCCTGGAGTGGTTGGGGAGC
#>                                                                                                            FR3.nt
#> 1 AATTACAATCCCTCCCTCAAGCGTCGAGTCACCATGTCAGTAGACACGGCCAGGAAGCAGTTGTCCCTGAAATTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTC
#> 2 TTCTACAACCCCTCCCTCAAAAGTCGAGTCACCATGTCAGTAGACACGTCCAAGAACCAGTTCTCCCTGAGGTTGACCTCTGTGACCGCTGCGGACACGGCCGTGTATTTC
#> 3 AACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCATTAGACACGTCCAAGAACCAGTTCTCCCTGAACCTGAGCTCTGTGACCGCTGCGGACACGGCCGTGTATTTT
#> 4 AACTACAACCCCTCCCTCAAGAGTCGAGTCAGCATATCCTTAGACACGTCCAAGAACCAGTTCTCCCTGACCCTGAGGTCTGTGACCGCTGCGGACACGGCCGTGTATTTT
#> 5 AGATACAGCCCGTCCTTCCAAGACCAGGTCACCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTAC
#> 6 AGATACAGCCCGTCCTTCCAAGACCAGGTCAGCATGTCAGTCGGCAAGTCCATCGGCACCGCCTACCTGCAGTGGAGCAGCCTGAAGGCCTCGGACACCGCCATGTATTAC
#> 7 AGATACAATCCGTCCTTCCAAGGCCAAGTCACCATCTCAGCCGACAGGTCCACCACCACCGCCTACCTGCAGTGGAGCAGTCTGAAGGCCTCGGACACCGCCATATATTAC
#> 8 AGATACAGTCCGTCCTTCCAAGGCCAGGTCACCATCTCAGCCGACAAGTCCCTCAGTACCGCCTACTTGCAGTGGAACAGCCTGCAGGCCTCGGACACCGCCATCTACTAC
#>                            FR4.nt
#> 1 GGCCAGGGAGCCCTGGTCACCGTCTCCTCAG
#> 2 GGCCCGGGAACCCTGGTCACCGTCTCCTCAG
#> 3 GGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#> 4 GGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#> 5 GGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#> 6 GGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#> 7 GGCCAGGGAGCCCTGGTCACCGTCTCCTCAG
#> 8 GGCCAGGGAACCCTGGTCACCGTCTCCTCAG
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             Germline.Alignment.V
#> 1 c, c, a, a, g, g, g, g, t, t, g, g, c, c, a, a, g, g, c, c, t, t, g, g, c, c, a, a, g, g, g, g, a, a, g, g, t, t, c, c, g, g, g, g, g, g, c, c, c, c, c, c, a, a, g, g, g, g, a, a, c, c, t, t, g, g, g, g, t, t, g, g, a, a, a, a, g, g, c, c, c, c, t, t, t, t, c, c, g, g, g, g, a, a, g, g, a, a, c, c, c, c, c, c, t, t, g, g, t, t, c, c, c, c, c, c, t, t, c, c, a, a, c, c, c, c, t, t, g, g, c, c, a, a, c, c, t, t, g, g, t, t, c, c, t, t, c, c, t, t, g, g, g, g, t, t, g, g, g, g, c, c, t, t, c, c, c, c, g, g, t, t, c, c, a, a, g, g, c, c, a, a, g, g, t, t, g, g, g, g, t, t, a, g, g, g, t, t, t, t, a, a, c, c, t, t, a, a, c, c, t, t, g, g, g, g, a, a, g, g, c, c, t, t, g, g, g, g, a, a, t, t, c, c, c, c, g, g, g, g, c, c, a, a, g, g, c, c, c, c, c, c, c, c, c, c, a, a, g, g, g, g, g, g, a, a, a, a, g, g, g, g, g, g, a, a, c, c, t, t, g, g, g, g, a, a, g, g, t, t, g, g, g, g, a, a, t, t, t, t, g, g, g, g, g, g, t, t, a, a, t, t, a, g, t, t, c, c, t, c, a, a, t, t, t, c, a, a, c, c, a, a, g, g, t, t, g, g, g, g, g, g, a, a, g, g, c, c, a, a, c, c, c, c, a, a, a, a, c, t, t, t, a, a, c, c, a, a, a, a, c, t, c, c, c, c, c, c, t, t, c, c, c, c, c, c, t, t, c, c, a, a, a, a, g, g, a, c, g, g, t, t, c, c, g, g, a, a, g, g, t, t, c, c, a, a, c, c, c, c, a, a, t, t, a, g, t, t, c, c, a, a, g, g, t, t, a, a, g, g, a, a, c, c, a, a, c, c, g, g, t, g, c, c, c, c, a, a, a, g, g, g, a, a, a, a, c, g, c, c, a, a, g, g, t, t, t, t, c, g, t, t, c, c, c, c, c, c, t, t, g, g, a, a, a, a, g, a, c, t, t, t, g, g, a, a, g, c, c, c, t, t, c, c, t, t, g, g, t, t, g, g, a, a, c, c, c, c, g, g, c, c, t, t, g, g, c, c, g, g, g, g, a, a, c, c, a, a, c, c, g, g, g, g, c, c, c, c, g, g, t, t, g, g, t, t, a, a, t, t, t, t, a, t, c, c
#> 2 c, c, a, a, g, g, g, g, t, t, g, g, c, c, a, a, g, g, c, c, t, t, g, g, c, c, a, a, g, g, g, g, a, a, g, g, t, t, c, c, g, g, g, g, g, g, c, c, c, c, c, c, a, a, g, g, g, g, a, a, c, c, t, t, g, g, g, g, t, t, g, g, a, a, a, a, g, g, c, c, c, c, t, t, t, t, c, c, g, g, g, g, a, a, g, g, a, a, c, c, c, c, c, c, t, t, g, g, t, t, c, c, c, c, c, c, t, t, c, c, a, a, c, c, c, c, t, t, g, g, c, c, a, a, c, c, t, t, g, g, t, t, c, c, t, t, c, c, t, t, g, g, g, g, t, t, g, g, g, g, c, c, t, t, c, c, c, c, g, g, t, t, c, c, a, a, g, g, c, c, a, a, g, g, t, t, g, g, g, g, t, t, a, g, g, g, t, t, t, t, a, a, c, c, t, t, a, a, c, c, t, t, g, g, g, g, a, a, g, g, c, c, t, t, g, g, g, g, a, a, t, t, c, c, c, c, g, g, g, g, c, c, a, a, g, g, c, c, c, c, c, c, c, c, c, c, a, a, g, g, g, g, g, g, a, a, a, a, g, g, g, g, g, g, a, a, c, c, t, t, g, g, g, g, a, a, g, g, t, t, g, g, g, g, a, a, t, t, t, t, g, g, g, g, g, a, t, t, a, a, t, t, a, a, t, t, c, t, t, c, a, a, t, t, t, t, a, a, c, c, a, a, g, g, t, t, g, g, g, g, g, g, a, a, g, a, c, c, a, t, c, c, c, c, a, t, a, t, c, c, t, t, a, a, c, c, a, a, a, a, c, c, c, c, c, c, c, c, t, t, c, c, c, c, c, c, t, t, c, c, a, a, a, a, g, a, a, a, g, g, t, t, c, c, g, g, a, a, g, g, t, t, c, c, a, a, c, c, c, c, a, a, t, t, a, g, t, t, c, c, a, a, g, g, t, t, a, a, g, g, a, a, c, c, a, a, c, c, g, g, t, t, c, c, c, c, a, a, a, a, g, g, a, a, a, a, c, c, c, c, a, a, g, g, t, t, t, t, c, c, t, t, c, c, c, c, c, c, t, t, g, g, a, a, a, g, g, g, c, t, t, t, g, g, a, a, g, c, c, c, t, t, c, c, t, t, g, g, t, t, g, g, a, a, c, c, c, c, g, g, c, c, t, t, g, g, c, c, g, g, g, g, a, a, c, c, a, a, c, c, g, g, g, g, c, c, c, c, g, g, t, t, g, g, t, t, a, a, t, t, t, t, a, t, c, c
#> 3 c, c, a, a, g, g, g, g, t, t, g, g, c, c, a, a, g, g, c, c, t, t, g, g, c, c, a, a, g, g, g, g, a, a, g, g, t, t, c, c, g, g, g, g, g, g, c, c, c, c, c, c, a, a, g, g, g, g, a, a, c, c, t, t, g, g, g, g, t, t, g, g, a, a, a, a, g, g, c, c, c, c, t, t, t, t, c, c, g, g, g, g, a, a, g, g, a, a, c, c, c, c, c, c, t, t, g, g, t, t, c, c, c, c, c, c, t, t, c, c, a, a, c, c, c, c, t, t, g, g, c, c, a, a, c, c, t, t, g, g, t, t, c, c, t, t, c, c, t, t, g, g, g, g, t, t, g, g, g, g, c, c, t, t, c, c, c, c, g, g, t, t, c, c, a, a, g, g, c, c, a, a, g, g, t, t, g, g, g, g, t, t, a, g, g, g, t, t, t, t, a, a, c, c, t, t, a, a, c, c, t, t, g, g, g, g, a, a, g, g, c, c, t, t, g, g, g, g, a, a, t, t, c, c, c, c, g, g, g, g, c, c, a, a, g, g, c, c, c, c, c, c, c, c, c, c, a, a, g, g, g, g, g, g, a, a, a, a, g, g, g, g, g, g, a, a, c, c, t, t, g, g, g, g, a, a, g, g, t, t, g, g, g, g, a, a, t, t, t, t, g, g, g, g, g, g, t, t, a, a, t, t, a, a, t, t, c, c, t, t, a, a, t, t, t, t, a, a, c, c, a, a, g, g, t, t, g, g, g, g, g, g, a, a, g, g, c, t, a, a, c, c, c, c, a, a, a, a, c, c, t, t, a, a, c, c, a, a, a, a, c, c, c, c, c, c, c, c, t, t, c, c, c, c, c, c, t, t, c, c, a, a, a, a, g, g, a, a, g, g, t, t, c, c, g, g, a, a, g, g, t, t, c, c, a, a, c, g, c, c, a, a, t, t, a, a, t, t, c, c, a, a, g, t, t, t, a, a, g, g, a, a, c, c, a, a, c, c, g, g, t, t, c, c, c, c, a, a, a, a, g, g, a, a, a, a, c, c, c, c, a, a, g, g, t, t, t, t, c, c, t, t, c, c, c, c, c, c, t, t, g, g, a, a, a, a, g, c, c, c, t, t, g, g, a, a, g, g, c, c, t, t, c, c, t, t, g, g, t, t, g, g, a, a, c, c, c, c, g, g, c, c, t, t, g, g, c, c, g, g, g, g, a, a, c, c, a, a, c, c, g, g, g, g, c, c, c, c, g, g, t, t, g, g, t, t, a, a, t, t, t, t, a, t, c, t
#> 4 c, c, a, a, g, g, g, g, t, t, g, g, c, c, a, a, g, g, c, c, t, t, g, g, c, c, a, a, g, g, g, g, a, a, g, g, t, t, c, c, g, g, g, g, g, g, c, c, c, c, c, c, a, a, g, g, g, g, a, a, c, c, t, t, g, g, g, g, t, t, g, g, a, a, a, a, g, g, c, c, c, c, t, t, t, t, c, c, g, g, g, g, a, a, g, g, a, a, c, c, c, c, c, c, t, t, g, g, t, t, c, c, c, c, c, c, t, t, c, c, a, a, c, c, c, c, t, t, g, g, c, c, a, a, c, c, t, t, g, g, t, t, c, c, t, t, c, c, t, t, g, g, g, g, t, t, g, g, g, g, c, c, t, t, c, c, c, c, g, g, t, t, c, c, a, a, g, g, c, c, a, a, g, g, t, t, g, g, g, g, t, t, a, g, g, g, t, t, t, t, a, a, c, c, t, t, a, a, c, c, t, t, g, g, g, g, a, a, g, g, c, c, t, t, g, g, g, g, a, a, t, t, c, c, c, c, g, g, g, g, c, c, a, a, g, g, c, c, c, c, c, c, c, c, c, c, a, a, g, g, g, g, g, g, a, a, a, a, g, g, g, g, g, g, a, a, c, c, t, t, g, g, g, g, a, a, g, g, t, t, g, g, g, g, a, a, t, t, t, t, g, g, g, g, g, g, t, a, a, a, t, t, a, a, t, t, c, c, t, t, a, a, t, t, t, t, a, a, c, c, a, a, g, g, t, t, g, g, g, g, g, g, a, a, g, c, c, t, a, a, c, c, c, c, a, a, a, a, c, c, t, t, a, a, c, c, a, a, a, a, c, c, c, c, c, c, c, c, t, t, c, c, c, c, c, c, t, t, c, c, a, a, a, a, g, g, a, a, g, g, t, t, c, c, g, g, a, a, g, g, t, t, c, c, a, a, c, g, c, c, a, a, t, t, a, a, t, t, c, c, a, c, g, t, t, t, a, a, g, g, a, a, c, c, a, a, c, c, g, g, t, t, c, c, c, c, a, a, a, a, g, g, a, a, a, a, c, c, c, c, a, a, g, g, t, t, t, t, c, c, t, t, c, c, c, c, c, c, t, t, g, g, a, a, a, c, g, c, c, c, t, t, g, g, a, a, g, g, c, g, t, t, c, c, t, t, g, g, t, t, g, g, a, a, c, c, c, c, g, g, c, c, t, t, g, g, c, c, g, g, g, g, a, a, c, c, a, a, c, c, g, g, g, g, c, c, c, c, g, g, t, t, g, g, t, t, a, a, t, t, t, t, a, t, c, t
#> 5                   g, g, a, a, g, g, g, g, t, t, g, g, c, c, a, a, g, g, c, c, t, t, g, g, g, g, t, t, g, g, c, c, a, a, g, g, t, t, c, c, t, t, g, g, g, g, a, a, g, g, c, c, a, a, g, g, a, a, g, g, g, g, t, t, g, g, a, a, a, a, a, a, a, a, a, a, g, g, c, c, c, c, c, c, g, g, g, g, g, g, g, g, a, a, g, g, t, t, c, c, t, t, c, c, t, t, g, g, a, a, a, a, g, g, a, a, t, t, c, c, t, t, c, c, c, c, t, t, g, g, t, t, a, a, a, a, g, g, g, g, g, g, t, t, t, t, c, c, t, t, g, g, g, g, a, a, t, t, a, a, c, c, a, a, g, g, c, c, t, t, t, t, t, t, a, a, c, c, c, c, a, a, g, g, c, c, t, t, a, a, c, c, t, t, g, g, g, g, a, a, t, t, c, c, g, g, g, g, c, c, t, t, g, g, g, g, g, g, t, t, g, g, c, c, g, g, c, c, c, c, a, a, g, g, a, a, t, t, g, g, c, c, c, c, c, c, g, g, g, g, g, g, a, a, a, a, a, a, g, g, g, g, c, c, c, c, t, t, g, g, g, g, a, a, g, g, t, t, g, g, g, g, a, a, t, t, g, g, g, g, g, g, g, g, a, a, t, t, c, c, a, a, t, t, c, c, t, t, a, a, t, t, c, c, c, c, t, t, g, g, g, g, t, t, g, g, a, a, c, c, t, t, c, c, t, t, g, g, a, a, t, t, a, a, c, c, c, c, a, a, g, g, a, a, t, t, a, a, c, c, a, a, g, g, c, c, c, c, c, c, g, g, t, t, c, c, c, c, t, t, t, t, c, c, c, c, a, a, a, a, g, g, g, a, c, c, c, c, a, a, g, g, g, g, t, t, c, c, a, a, c, c, c, c, a, a, t, t, c, g, t, t, c, c, a, a, g, g, c, t, c, c, g, g, a, g, c, c, a, a, a, a, g, g, t, t, c, c, c, c, a, a, t, t, c, c, a, g, g, g, c, c, a, a, c, c, c, c, g, g, c, c, c, c, t, t, a, a, c, c, c, c, t, t, g, g, c, c, a, a, g, g, t, t, g, g, g, g, a, a, g, g, c, c, a, a, g, g, c, c, c, c, t, t, g, g, a, a, a, a, g, g, g, g, c, c, c, c, t, t, c, c, g, g, g, g, a, a, c, c, a, a, c, c, c, c, g, g, c, c, c, c, a, a, t, t, g, g, t, t, a, a, t, t, t, t, a, a, c, c
#> 6                   g, g, a, a, g, g, g, g, t, t, g, g, c, c, a, a, g, g, c, c, t, t, g, g, g, g, t, t, g, g, c, c, a, a, g, g, t, t, c, c, t, t, g, g, g, g, a, a, g, g, c, c, a, a, g, g, a, a, g, g, g, g, t, t, g, g, a, a, a, a, a, a, a, a, a, a, g, g, c, c, c, c, c, c, g, g, g, g, g, g, g, g, a, a, g, g, t, t, c, c, t, t, c, c, t, t, g, g, a, a, a, a, g, g, a, a, t, t, c, c, t, t, c, c, c, c, t, t, g, g, t, t, a, a, a, a, g, g, g, g, g, g, t, t, t, t, c, c, t, t, g, g, g, g, a, a, t, t, a, a, c, c, a, a, g, g, c, c, t, t, t, t, t, t, a, a, c, c, c, c, a, a, g, g, c, c, t, t, a, a, c, c, t, t, g, g, g, g, a, a, t, t, c, c, g, g, g, g, c, c, t, t, g, g, g, g, g, g, t, t, g, g, c, c, g, g, c, c, c, c, a, a, g, g, a, a, t, t, g, g, c, c, c, c, c, c, g, g, g, g, g, g, a, a, a, a, a, a, g, g, g, g, c, c, c, c, t, t, g, g, g, g, a, a, g, g, t, t, g, g, g, g, a, a, t, t, g, g, g, g, g, g, g, a, a, a, t, t, c, c, a, a, t, t, c, c, t, t, a, a, t, t, c, c, c, c, t, t, g, g, g, g, t, t, g, g, a, a, c, c, t, t, c, c, t, t, g, g, a, a, t, t, a, a, c, c, c, c, a, a, g, g, a, a, t, t, a, a, c, c, a, a, g, g, c, c, c, c, c, c, g, g, t, t, c, c, c, c, t, t, t, t, c, c, c, c, a, a, a, a, g, g, g, a, c, c, c, c, a, a, g, g, g, g, t, t, c, c, a, a, c, g, c, c, a, a, t, t, c, g, t, t, c, c, a, a, g, g, c, t, c, c, g, g, a, g, c, c, a, a, a, a, g, g, t, t, c, c, c, c, a, a, t, t, c, c, a, g, g, g, c, c, a, a, c, c, c, c, g, g, c, c, c, c, t, t, a, a, c, c, c, c, t, t, g, g, c, c, a, a, g, g, t, t, g, g, g, g, a, a, g, g, c, c, a, a, g, g, c, c, c, c, t, t, g, g, a, a, a, a, g, g, g, g, c, c, c, c, t, t, c, c, g, g, g, g, a, a, c, c, a, a, c, c, c, c, g, g, c, c, c, c, a, a, t, t, g, g, t, t, a, a, t, t, t, t, a, a, c, c
#> 7                   g, g, a, a, g, g, g, g, t, t, g, g, c, c, a, a, g, g, c, c, t, t, g, g, g, g, t, t, g, g, c, c, a, a, g, g, t, t, c, c, t, t, g, g, g, g, a, a, g, g, c, c, a, a, g, g, a, a, g, g, g, g, t, t, g, g, a, a, a, a, a, a, a, a, a, a, g, g, c, c, c, c, c, c, g, g, g, g, g, g, g, g, a, a, g, g, t, t, c, c, t, t, c, c, t, t, g, g, a, a, a, a, g, g, a, a, t, t, c, c, t, t, c, c, c, c, t, t, g, g, t, t, a, a, a, a, g, g, g, g, g, g, t, t, t, t, c, c, t, t, g, g, g, g, a, a, t, t, a, a, c, c, a, a, g, g, c, c, t, t, t, t, t, t, a, a, c, c, c, c, a, a, g, g, c, c, t, t, a, a, c, c, t, t, g, g, g, g, a, a, t, t, c, c, g, g, g, g, c, c, t, t, g, g, g, g, g, g, t, t, g, g, c, c, g, g, c, c, c, c, a, a, g, g, a, a, t, t, g, g, c, c, c, c, c, c, g, g, g, g, g, a, a, a, a, a, a, a, g, g, g, g, c, c, c, c, t, t, g, g, g, g, a, a, g, g, t, t, g, g, g, g, a, g, t, t, g, g, g, g, g, g, g, a, a, t, t, c, c, g, a, a, t, t, c, c, t, t, a, a, t, t, c, c, c, c, t, t, g, c, g, g, t, t, g, g, a, a, c, c, t, t, c, c, t, t, g, g, a, a, t, t, a, a, c, t, c, c, a, a, g, g, a, a, t, t, a, a, c, c, a, a, g, a, c, t, c, c, c, c, g, g, t, t, c, c, c, c, t, t, t, t, c, c, c, c, a, a, a, a, g, g, g, g, c, c, c, c, a, a, g, a, g, g, t, t, c, c, a, a, c, c, c, c, a, a, t, t, c, c, t, t, c, c, a, a, g, g, c, c, c, c, g, g, a, a, c, c, a, a, a, g, g, g, t, t, c, c, c, c, a, a, t, c, c, c, a, a, g, c, c, c, a, a, c, c, c, c, g, g, c, c, c, c, t, t, a, a, c, c, c, c, t, t, g, g, c, c, a, a, g, g, t, t, g, g, g, g, a, a, g, g, c, c, a, a, g, g, c, t, c, c, t, t, g, g, a, a, a, a, g, g, g, g, c, c, c, c, t, t, c, c, g, g, g, g, a, a, c, c, a, a, c, c, c, c, g, g, c, c, c, c, a, a, t, t, g, a, t, t, a, a, t, t, t, t, a, a, c, c
#> 8                   g, g, a, a, g, g, g, g, t, t, g, g, c, c, a, a, g, g, c, c, t, t, g, g, g, g, t, t, g, g, c, c, a, a, g, g, t, t, c, c, t, t, g, g, g, g, a, a, g, g, c, c, a, a, g, g, a, a, g, g, g, g, t, t, g, g, a, a, a, a, a, a, a, a, a, a, g, g, c, c, c, c, c, c, g, g, g, g, g, g, g, g, a, a, g, g, t, t, c, c, t, t, c, c, t, t, g, g, a, a, a, a, g, g, a, a, t, t, c, c, t, t, c, c, c, c, t, t, g, g, t, t, a, a, a, a, g, g, g, g, g, g, t, t, t, t, c, c, t, t, g, g, g, g, a, a, t, t, a, a, c, c, a, a, g, g, c, c, t, t, t, t, t, t, a, a, c, c, c, c, a, a, g, g, c, c, t, t, a, a, c, c, t, t, g, g, g, g, a, a, t, t, c, c, g, g, g, g, c, c, t, t, g, g, g, g, g, g, t, t, g, g, c, c, g, g, c, c, c, c, a, a, g, g, a, a, t, t, g, g, c, c, c, c, c, c, g, g, g, g, g, g, a, a, a, a, a, a, g, g, g, g, c, c, c, c, t, t, g, g, g, g, a, a, g, g, t, t, g, g, g, g, a, t, t, t, g, g, g, g, g, g, g, g, a, a, t, g, c, c, a, a, t, t, c, c, t, t, a, a, t, c, c, c, c, c, t, t, g, c, g, g, t, t, g, g, a, a, c, c, t, t, c, c, t, t, g, g, a, a, t, c, a, a, c, c, c, t, a, a, g, g, a, a, t, t, a, a, c, c, a, a, g, g, c, t, c, c, c, c, g, g, t, t, c, c, c, c, t, t, t, t, c, c, c, c, a, a, a, a, g, g, g, g, c, c, c, c, a, a, g, g, g, g, t, t, c, c, a, a, c, c, c, c, a, a, t, t, c, c, t, t, c, c, a, a, g, g, c, c, c, c, g, g, a, a, c, c, a, a, a, a, g, g, t, t, c, c, c, c, a, c, t, t, c, c, a, a, g, g, c, t, a, a, c, c, c, c, g, g, c, c, c, c, t, t, a, a, c, c, c, t, t, t, g, g, c, c, a, a, g, g, t, t, g, g, g, g, a, a, g, a, c, c, a, a, g, g, c, c, c, c, t, t, g, g, a, c, a, a, g, g, g, g, c, c, c, c, t, t, c, c, g, g, g, g, a, a, c, c, a, a, c, c, c, c, g, g, c, c, c, c, a, a, t, t, g, c, t, t, a, a, t, c, t, t, a, a, c, c
#>                                                                                                                                                                       Germline.Alignment.J
#> 1 g, g, g, g, c, c, c, c, a, a, a, g, g, g, g, g, a, a, a, g, c, c, c, c, c, c, t, t, g, g, g, g, t, t, c, c, a, a, c, c, c, c, g, g, t, t, c, c, t, t, c, c, c, c, t, t, c, c, a, a, g, g
#> 2 g, g, g, g, c, c, c, c, a, c, a, g, g, g, g, g, a, a, a, a, c, c, c, c, c, c, t, t, g, g, g, g, t, t, c, c, a, a, c, c, c, c, g, g, t, t, c, c, t, t, c, c, c, c, t, t, c, c, a, a, g, g
#> 3 g, g, g, g, c, c, c, c, a, a, a, g, g, g, g, g, a, a, a, a, c, c, c, c, c, c, t, t, g, g, g, g, t, t, c, c, a, a, c, c, c, c, g, g, t, t, c, c, t, t, c, c, c, c, t, t, c, c, a, a, g, g
#> 4 g, g, g, g, c, c, c, c, a, a, a, g, g, g, g, g, a, a, a, a, c, c, c, c, c, c, t, t, g, g, g, g, t, t, c, c, a, a, c, c, c, c, g, g, t, t, c, c, t, t, c, c, c, c, t, t, c, c, a, a, g, g
#> 5 g, g, g, g, c, c, c, c, a, a, a, g, g, g, g, g, a, a, a, a, c, c, c, c, c, c, t, t, g, g, g, g, t, t, c, c, a, a, c, c, c, c, g, g, t, t, c, c, t, t, c, c, c, c, t, t, c, c, a, a, g, g
#> 6 g, g, g, g, c, c, c, c, a, a, a, g, g, g, g, g, a, a, a, a, c, c, c, c, c, c, t, t, g, g, g, g, t, t, c, c, a, a, c, c, c, c, g, g, t, t, c, c, t, t, c, c, c, c, t, t, c, c, a, a, g, g
#> 7 g, g, g, g, c, c, c, c, a, a, a, g, g, g, g, g, a, a, a, g, c, c, c, c, c, c, t, t, g, g, g, g, t, t, c, c, a, a, c, c, c, c, g, g, t, t, c, c, t, t, c, c, c, c, t, t, c, c, a, a, g, g
#> 8 g, g, g, g, c, c, c, c, a, a, a, g, g, g, g, g, a, a, a, a, c, c, c, c, c, c, t, t, g, g, g, g, t, t, c, c, a, a, c, c, c, c, g, g, t, t, c, c, t, t, c, c, c, c, t, t, c, c, a, a, g, g
#>   Substitutions Insertions Deletions Mutations
#> 1            18          0         0        18
#> 2            16          0         0        16
#> 3             8          0         0         8
#> 4            13          0         0        13
#> 5             6          0         0         6
#> 6             8          0         0         8
#> 7            18          0         0        18
#> 8            15          0         0        15
#>