Visualisation of distributions using boxplots

Visualisation of distributions using ggplot2-based boxplots.

Arguments

.data
Input matrix or data frame.

.by
Pass NA if you want to plot samples without grouping.

You can pass a character vector with one or several column names from ".meta" to group your data before plotting. In this case you should provide ".meta".

You can pass a character vector that exactly matches the number of samples in your data, each value should correspond to a sample's property. It will be used to group data based on the values provided. Note that in this case you should pass NA to ".meta".

.meta
A metadata object. An R dataframe with sample names and their properties, such as, age, serostatus or hla.

.title
The text for the title of the plot.

.ncol
Number of columns to display. Provide NA (by default) if you want the function to automatically detect the optimal number of columns.

.labs
Character vector of length two with names for x-axis and y-axis, respectively.

.melt
If TRUE then apply melt to the ".data" before plotting. In this case ".data" is supposed to be a data frame with the first character column reserved for names of genes and other numeric columns reserved to counts or frequencies of genes. Each numeric column should be associated with a specific repertoire sample.

.legend
If TRUE then plots the legend. If FALSE removes the legend from the plot. If NA automatically detects the best way to display legends.

See also